split up the creep api files
