+++ docs/assets/js/search.js
@@ -1,3 +1,3 @@
 var typedoc = typedoc || {};
             typedoc.search = typedoc.search || {};
-            typedoc.search.data = {"kinds":{"1":"External module","32":"Variable","64":"Function","128":"Class","1024":"Property","2048":"Method","65536":"Type literal","262144":"Accessor","2097152":"Object literal"},"rows":[{"id":0,"kind":1,"name":"\"src/utils/config\"","url":"modules/_src_utils_config_.html","classes":"tsd-kind-external-module"},{"id":1,"kind":32,"name":"NO_CACHING_MEMORY","url":"modules/_src_utils_config_.html#no_caching_memory","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"src/utils/config\""},{"id":2,"kind":32,"name":"ALLOW_CUSTOM_ERRORS","url":"modules/_src_utils_config_.html#allow_custom_errors","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"src/utils/config\""},{"id":3,"kind":1,"name":"\"src/Helpers/UtilHelper\"","url":"modules/_src_helpers_utilhelper_.html","classes":"tsd-kind-external-module"},{"id":4,"kind":128,"name":"UtilHelper","url":"classes/_src_helpers_utilhelper_.utilhelper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/UtilHelper\""},{"id":5,"kind":2048,"name":"throwError","url":"classes/_src_helpers_utilhelper_.utilhelper.html#throwerror","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/UtilHelper\".UtilHelper"},{"id":6,"kind":1,"name":"\"src/Helpers/Normalize\"","url":"modules/_src_helpers_normalize_.html","classes":"tsd-kind-external-module"},{"id":7,"kind":128,"name":"Normalize","url":"classes/_src_helpers_normalize_.normalize.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/Normalize\""},{"id":8,"kind":2048,"name":"roomName","url":"classes/_src_helpers_normalize_.normalize.html#roomname","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/Normalize\".Normalize"},{"id":9,"kind":2048,"name":"roomObject","url":"classes/_src_helpers_normalize_.normalize.html#roomobject","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/Normalize\".Normalize"},{"id":10,"kind":2048,"name":"roomPos","url":"classes/_src_helpers_normalize_.normalize.html#roompos","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/Normalize\".Normalize"},{"id":11,"kind":2048,"name":"creepObject","url":"classes/_src_helpers_normalize_.normalize.html#creepobject","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/Normalize\".Normalize"},{"id":12,"kind":1,"name":"\"src/Helpers/MemoryHelper_Room\"","url":"modules/_src_helpers_memoryhelper_room_.html","classes":"tsd-kind-external-module"},{"id":13,"kind":128,"name":"MemoryHelper_Room","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/MemoryHelper_Room\""},{"id":14,"kind":2048,"name":"updateRoomMemory","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updateroommemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":15,"kind":2048,"name":"updateDependentRooms","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatedependentrooms","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":16,"kind":2048,"name":"updateHostileCreeps","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatehostilecreeps","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":17,"kind":2048,"name":"updateMyCreeps","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatemycreeps","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":18,"kind":2048,"name":"updateConstructionSites","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updateconstructionsites","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":19,"kind":2048,"name":"updateStructures","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatestructures","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":20,"kind":2048,"name":"updateSources","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatesources","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":21,"kind":2048,"name":"updateRoomState","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updateroomstate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":22,"kind":2048,"name":"updateDefcon","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatedefcon","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":23,"kind":2048,"name":"updateDomesticLimits","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatedomesticlimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":24,"kind":2048,"name":"updateRemoteLimits","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updateremotelimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":25,"kind":2048,"name":"updateMilitaryLimits","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatemilitarylimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":26,"kind":1,"name":"\"src/Api/Memory.Api\"","url":"modules/_src_api_memory_api_.html","classes":"tsd-kind-external-module"},{"id":27,"kind":128,"name":"MemoryApi","url":"classes/_src_api_memory_api_.memoryapi.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Api/Memory.Api\""},{"id":28,"kind":2048,"name":"garbageCollection","url":"classes/_src_api_memory_api_.memoryapi.html#garbagecollection","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":29,"kind":2048,"name":"initRoomMemory","url":"classes/_src_api_memory_api_.memoryapi.html#initroommemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":30,"kind":2048,"name":"getRoomMemory","url":"classes/_src_api_memory_api_.memoryapi.html#getroommemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":31,"kind":2048,"name":"initCreepMemory","url":"classes/_src_api_memory_api_.memoryapi.html#initcreepmemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":32,"kind":2048,"name":"getMyCreeps","url":"classes/_src_api_memory_api_.memoryapi.html#getmycreeps","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":33,"kind":2048,"name":"getHostileCreeps","url":"classes/_src_api_memory_api_.memoryapi.html#gethostilecreeps","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":34,"kind":2048,"name":"getStructures","url":"classes/_src_api_memory_api_.memoryapi.html#getstructures","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":35,"kind":2048,"name":"getStructureOfType","url":"classes/_src_api_memory_api_.memoryapi.html#getstructureoftype","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":36,"kind":2048,"name":"getConstructionSites","url":"classes/_src_api_memory_api_.memoryapi.html#getconstructionsites","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":37,"kind":2048,"name":"getSources","url":"classes/_src_api_memory_api_.memoryapi.html#getsources","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":38,"kind":2048,"name":"getRemoteRooms","url":"classes/_src_api_memory_api_.memoryapi.html#getremoterooms","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":39,"kind":2048,"name":"getClaimRooms","url":"classes/_src_api_memory_api_.memoryapi.html#getclaimrooms","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":40,"kind":2048,"name":"getAttackRooms","url":"classes/_src_api_memory_api_.memoryapi.html#getattackrooms","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":41,"kind":1,"name":"\"src/Helpers/RoomHelper\"","url":"modules/_src_helpers_roomhelper_.html","classes":"tsd-kind-external-module"},{"id":42,"kind":128,"name":"RoomHelper","url":"classes/_src_helpers_roomhelper_.roomhelper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/RoomHelper\""},{"id":43,"kind":2048,"name":"isOwnedRoom","url":"classes/_src_helpers_roomhelper_.roomhelper.html#isownedroom","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":44,"kind":2048,"name":"isAllyRoom","url":"classes/_src_helpers_roomhelper_.roomhelper.html#isallyroom","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":45,"kind":2048,"name":"isSourceKeeperRoom","url":"classes/_src_helpers_roomhelper_.roomhelper.html#issourcekeeperroom","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":46,"kind":2048,"name":"isHighwayRoom","url":"classes/_src_helpers_roomhelper_.roomhelper.html#ishighwayroom","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":47,"kind":2048,"name":"inTravelRange","url":"classes/_src_helpers_roomhelper_.roomhelper.html#intravelrange","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":48,"kind":2048,"name":"isExistInRoom","url":"classes/_src_helpers_roomhelper_.roomhelper.html#isexistinroom","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":49,"kind":2048,"name":"getStoredAmount","url":"classes/_src_helpers_roomhelper_.roomhelper.html#getstoredamount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":50,"kind":2048,"name":"getStoredCapacity","url":"classes/_src_helpers_roomhelper_.roomhelper.html#getstoredcapacity","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":51,"kind":2048,"name":"getTowerDamageAtRange","url":"classes/_src_helpers_roomhelper_.roomhelper.html#gettowerdamageatrange","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":52,"kind":2048,"name":"excecuteEveryTicks","url":"classes/_src_helpers_roomhelper_.roomhelper.html#excecuteeveryticks","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":53,"kind":2048,"name":"isContainerMining","url":"classes/_src_helpers_roomhelper_.roomhelper.html#iscontainermining","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":54,"kind":2048,"name":"isUpgraderLink","url":"classes/_src_helpers_roomhelper_.roomhelper.html#isupgraderlink","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":55,"kind":2048,"name":"chooseTowerTarget","url":"classes/_src_helpers_roomhelper_.roomhelper.html#choosetowertarget","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":56,"kind":2048,"name":"getWallLevelDifference","url":"classes/_src_helpers_roomhelper_.roomhelper.html#getwallleveldifference","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":57,"kind":2048,"name":"numHostileCreeps","url":"classes/_src_helpers_roomhelper_.roomhelper.html#numhostilecreeps","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":58,"kind":2048,"name":"numRemoteRooms","url":"classes/_src_helpers_roomhelper_.roomhelper.html#numremoterooms","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":59,"kind":2048,"name":"numClaimRooms","url":"classes/_src_helpers_roomhelper_.roomhelper.html#numclaimrooms","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":60,"kind":2048,"name":"numAttackRooms","url":"classes/_src_helpers_roomhelper_.roomhelper.html#numattackrooms","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":61,"kind":2048,"name":"numSources","url":"classes/_src_helpers_roomhelper_.roomhelper.html#numsources","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":62,"kind":2048,"name":"numRemoteSources","url":"classes/_src_helpers_roomhelper_.roomhelper.html#numremotesources","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":63,"kind":2048,"name":"numRemoteDefenders","url":"classes/_src_helpers_roomhelper_.roomhelper.html#numremotedefenders","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":64,"kind":1,"name":"\"src/Helpers/ConsoleCommands\"","url":"modules/_src_helpers_consolecommands_.html","classes":"tsd-kind-external-module"},{"id":65,"kind":2097152,"name":"ConsoleCommands","url":"modules/_src_helpers_consolecommands_.html#consolecommands","classes":"tsd-kind-object-literal tsd-parent-kind-external-module","parent":"\"src/Helpers/ConsoleCommands\""},{"id":66,"kind":64,"name":"removeConstructionSites","url":"modules/_src_helpers_consolecommands_.html#consolecommands.removeconstructionsites","classes":"tsd-kind-function tsd-parent-kind-object-literal","parent":"\"src/Helpers/ConsoleCommands\".ConsoleCommands"},{"id":67,"kind":64,"name":"removeFlags","url":"modules/_src_helpers_consolecommands_.html#consolecommands.removeflags","classes":"tsd-kind-function tsd-parent-kind-object-literal","parent":"\"src/Helpers/ConsoleCommands\".ConsoleCommands"},{"id":68,"kind":64,"name":"displayRoomStatus","url":"modules/_src_helpers_consolecommands_.html#consolecommands.displayroomstatus","classes":"tsd-kind-function tsd-parent-kind-object-literal","parent":"\"src/Helpers/ConsoleCommands\".ConsoleCommands"},{"id":69,"kind":64,"name":"killAllCreeps","url":"modules/_src_helpers_consolecommands_.html#consolecommands.killallcreeps","classes":"tsd-kind-function tsd-parent-kind-object-literal","parent":"\"src/Helpers/ConsoleCommands\".ConsoleCommands"},{"id":70,"kind":64,"name":"sendResource","url":"modules/_src_helpers_consolecommands_.html#consolecommands.sendresource","classes":"tsd-kind-function tsd-parent-kind-object-literal","parent":"\"src/Helpers/ConsoleCommands\".ConsoleCommands"},{"id":71,"kind":1,"name":"\"src/Managers/EmpireManager\"","url":"modules/_src_managers_empiremanager_.html","classes":"tsd-kind-external-module"},{"id":72,"kind":128,"name":"EmpireManager","url":"classes/_src_managers_empiremanager_.empiremanager.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Managers/EmpireManager\""},{"id":73,"kind":2048,"name":"runEmpireManager","url":"classes/_src_managers_empiremanager_.empiremanager.html#runempiremanager","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Managers/EmpireManager\".EmpireManager"},{"id":74,"kind":1,"name":"\"src/Managers/MemoryManagement\"","url":"modules/_src_managers_memorymanagement_.html","classes":"tsd-kind-external-module"},{"id":75,"kind":128,"name":"MemoryManager","url":"classes/_src_managers_memorymanagement_.memorymanager.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Managers/MemoryManagement\""},{"id":76,"kind":2048,"name":"runMemoryManager","url":"classes/_src_managers_memorymanagement_.memorymanager.html#runmemorymanager","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Managers/MemoryManagement\".MemoryManager"},{"id":77,"kind":1,"name":"\"src/Managers/RoomManager\"","url":"modules/_src_managers_roommanager_.html","classes":"tsd-kind-external-module"},{"id":78,"kind":128,"name":"RoomManager","url":"classes/_src_managers_roommanager_.roommanager.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Managers/RoomManager\""},{"id":79,"kind":2048,"name":"runRoomManager","url":"classes/_src_managers_roommanager_.roommanager.html#runroommanager","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Managers/RoomManager\".RoomManager"},{"id":80,"kind":1,"name":"\"src/Managers/SpawnManager\"","url":"modules/_src_managers_spawnmanager_.html","classes":"tsd-kind-external-module"},{"id":81,"kind":128,"name":"SpawnManager","url":"classes/_src_managers_spawnmanager_.spawnmanager.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Managers/SpawnManager\""},{"id":82,"kind":2048,"name":"runSpawnManager","url":"classes/_src_managers_spawnmanager_.spawnmanager.html#runspawnmanager","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Managers/SpawnManager\".SpawnManager"},{"id":83,"kind":1,"name":"\"src/utils/ErrorMapper\"","url":"modules/_src_utils_errormapper_.html","classes":"tsd-kind-external-module"},{"id":84,"kind":128,"name":"ErrorMapper","url":"classes/_src_utils_errormapper_.errormapper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/utils/ErrorMapper\""},{"id":85,"kind":1024,"name":"_consumer","url":"classes/_src_utils_errormapper_.errormapper.html#_consumer","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-private tsd-is-static","parent":"\"src/utils/ErrorMapper\".ErrorMapper"},{"id":86,"kind":262144,"name":"consumer","url":"classes/_src_utils_errormapper_.errormapper.html#consumer","classes":"tsd-kind-get-signature tsd-parent-kind-class tsd-is-static","parent":"\"src/utils/ErrorMapper\".ErrorMapper"},{"id":87,"kind":1024,"name":"cache","url":"classes/_src_utils_errormapper_.errormapper.html#cache","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-static","parent":"\"src/utils/ErrorMapper\".ErrorMapper"},{"id":88,"kind":65536,"name":"__type","url":"classes/_src_utils_errormapper_.errormapper.html#cache.__type","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"\"src/utils/ErrorMapper\".ErrorMapper.cache"},{"id":89,"kind":2048,"name":"sourceMappedStackTrace","url":"classes/_src_utils_errormapper_.errormapper.html#sourcemappedstacktrace","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/utils/ErrorMapper\".ErrorMapper"},{"id":90,"kind":2048,"name":"wrapLoop","url":"classes/_src_utils_errormapper_.errormapper.html#wraploop","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/utils/ErrorMapper\".ErrorMapper"},{"id":91,"kind":1,"name":"\"src/main\"","url":"modules/_src_main_.html","classes":"tsd-kind-external-module"},{"id":92,"kind":32,"name":"loop","url":"modules/_src_main_.html#loop","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"src/main\""},{"id":93,"kind":65536,"name":"__type","url":"modules/_src_main_.html#loop.__type","classes":"tsd-kind-type-literal tsd-parent-kind-variable tsd-is-not-exported","parent":"\"src/main\".loop"},{"id":94,"kind":1,"name":"\"src/Api/CreepDomestic.Api\"","url":"modules/_src_api_creepdomestic_api_.html","classes":"tsd-kind-external-module"},{"id":95,"kind":128,"name":"CreepDomestic","url":"classes/_src_api_creepdomestic_api_.creepdomestic.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Api/CreepDomestic.Api\""},{"id":96,"kind":1,"name":"\"src/Api/CreepMili.Api\"","url":"modules/_src_api_creepmili_api_.html","classes":"tsd-kind-external-module"},{"id":97,"kind":128,"name":"CreepMili","url":"classes/_src_api_creepmili_api_.creepmili.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Api/CreepMili.Api\""},{"id":98,"kind":1,"name":"\"src/Api/CreepRemote.Api\"","url":"modules/_src_api_creepremote_api_.html","classes":"tsd-kind-external-module"},{"id":99,"kind":128,"name":"CreepRemote","url":"classes/_src_api_creepremote_api_.creepremote.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Api/CreepRemote.Api\""},{"id":100,"kind":1,"name":"\"src/Api/Empire.Api\"","url":"modules/_src_api_empire_api_.html","classes":"tsd-kind-external-module"},{"id":101,"kind":128,"name":"Empire","url":"classes/_src_api_empire_api_.empire.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Api/Empire.Api\""},{"id":102,"kind":2048,"name":"processNewFlags","url":"classes/_src_api_empire_api_.empire.html#processnewflags","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":103,"kind":2048,"name":"removeFlagMemory","url":"classes/_src_api_empire_api_.empire.html#removeflagmemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":104,"kind":2048,"name":"processNewRemoteFlag","url":"classes/_src_api_empire_api_.empire.html#processnewremoteflag","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":105,"kind":2048,"name":"processNewAttackFlag","url":"classes/_src_api_empire_api_.empire.html#processnewattackflag","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":106,"kind":2048,"name":"processNewClaimFlag","url":"classes/_src_api_empire_api_.empire.html#processnewclaimflag","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":107,"kind":2048,"name":"cleanDeadFlags","url":"classes/_src_api_empire_api_.empire.html#cleandeadflags","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":108,"kind":2048,"name":"findDependentRoom","url":"classes/_src_api_empire_api_.empire.html#finddependentroom","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":109,"kind":2048,"name":"findDependentRoomAuto","url":"classes/_src_api_empire_api_.empire.html#finddependentroomauto","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-private tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":110,"kind":2048,"name":"findDependentRoomManual","url":"classes/_src_api_empire_api_.empire.html#finddependentroommanual","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-private tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":111,"kind":1,"name":"\"src/Helpers/MemoryHelper\"","url":"modules/_src_helpers_memoryhelper_.html","classes":"tsd-kind-external-module"},{"id":112,"kind":128,"name":"MemoryHelper","url":"classes/_src_helpers_memoryhelper_.memoryhelper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/MemoryHelper\""},{"id":113,"kind":2048,"name":"getStructureOfType","url":"classes/_src_helpers_memoryhelper_.memoryhelper.html#getstructureoftype","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper\".MemoryHelper"},{"id":114,"kind":2048,"name":"getCreepOfRole","url":"classes/_src_helpers_memoryhelper_.memoryhelper.html#getcreepofrole","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper\".MemoryHelper"},{"id":115,"kind":2048,"name":"clearCreepMemory","url":"classes/_src_helpers_memoryhelper_.memoryhelper.html#clearcreepmemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper\".MemoryHelper"},{"id":116,"kind":2048,"name":"clearRoomMemory","url":"classes/_src_helpers_memoryhelper_.memoryhelper.html#clearroommemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper\".MemoryHelper"},{"id":117,"kind":1,"name":"\"src/Api/Room.Api\"","url":"modules/_src_api_room_api_.html","classes":"tsd-kind-external-module"},{"id":118,"kind":128,"name":"RoomApi","url":"classes/_src_api_room_api_.roomapi.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Api/Room.Api\""},{"id":119,"kind":2048,"name":"isHostilesInRoom","url":"classes/_src_api_room_api_.roomapi.html#ishostilesinroom","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":120,"kind":2048,"name":"setRoomState","url":"classes/_src_api_room_api_.roomapi.html#setroomstate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":121,"kind":2048,"name":"runTowers","url":"classes/_src_api_room_api_.roomapi.html#runtowers","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":122,"kind":2048,"name":"setDefconLevel","url":"classes/_src_api_room_api_.roomapi.html#setdefconlevel","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":123,"kind":2048,"name":"getRepairTargets","url":"classes/_src_api_room_api_.roomapi.html#getrepairtargets","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":124,"kind":2048,"name":"getExtensionsNeedFilled","url":"classes/_src_api_room_api_.roomapi.html#getextensionsneedfilled","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":125,"kind":2048,"name":"getTowersNeedFilled","url":"classes/_src_api_room_api_.roomapi.html#gettowersneedfilled","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":126,"kind":2048,"name":"getWallRepairTargets","url":"classes/_src_api_room_api_.roomapi.html#getwallrepairtargets","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":127,"kind":2048,"name":"getOpenSources","url":"classes/_src_api_room_api_.roomapi.html#getopensources","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":128,"kind":2048,"name":"getMiningContainer","url":"classes/_src_api_room_api_.roomapi.html#getminingcontainer","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":129,"kind":2048,"name":"isFull","url":"classes/_src_api_room_api_.roomapi.html#isfull","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":130,"kind":2048,"name":"getWallHpLimit","url":"classes/_src_api_room_api_.roomapi.html#getwallhplimit","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-private tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":131,"kind":2048,"name":"createJobQueue","url":"classes/_src_api_room_api_.roomapi.html#createjobqueue","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":132,"kind":2048,"name":"createJobQueueWorker","url":"classes/_src_api_room_api_.roomapi.html#createjobqueueworker","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":133,"kind":2048,"name":"createJobQueueLorry","url":"classes/_src_api_room_api_.roomapi.html#createjobqueuelorry","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":134,"kind":2048,"name":"createEnergyQueue","url":"classes/_src_api_room_api_.roomapi.html#createenergyqueue","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":135,"kind":2048,"name":"runLinks","url":"classes/_src_api_room_api_.roomapi.html#runlinks","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":136,"kind":2048,"name":"runTerminal","url":"classes/_src_api_room_api_.roomapi.html#runterminal","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":137,"kind":2048,"name":"runLabs","url":"classes/_src_api_room_api_.roomapi.html#runlabs","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":138,"kind":1,"name":"\"src/Helpers/SpawnHelper\"","url":"modules/_src_helpers_spawnhelper_.html","classes":"tsd-kind-external-module"},{"id":139,"kind":128,"name":"SpawnHelper","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/SpawnHelper\""},{"id":140,"kind":2048,"name":"verifyDescriptor","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#verifydescriptor","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":141,"kind":2048,"name":"generateParts","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateparts","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":142,"kind":2048,"name":"getBody_Grouped","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#getbody_grouped","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":143,"kind":2048,"name":"getBody_Collated","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#getbody_collated","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":144,"kind":2048,"name":"generateCreepName","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatecreepname","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":145,"kind":2048,"name":"generateMinerBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateminerbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":146,"kind":2048,"name":"generateMinerOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatemineroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":147,"kind":2048,"name":"generateHarvesterBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateharvesterbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":148,"kind":2048,"name":"generateHarvesterOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateharvesteroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":149,"kind":2048,"name":"generateWorkerBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateworkerbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":150,"kind":2048,"name":"generateWorkerOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateworkeroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":151,"kind":2048,"name":"generateLorryBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatelorrybody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":152,"kind":2048,"name":"generateLorryOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatelorryoptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":153,"kind":2048,"name":"generatePowerUpgraderBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatepowerupgraderbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":154,"kind":2048,"name":"generatePowerUpgraderOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatepowerupgraderoptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":155,"kind":2048,"name":"generateRemoteMinerBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremoteminerbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":156,"kind":2048,"name":"generateRemoteMinerOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremotemineroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":157,"kind":2048,"name":"generateRemoteHarvesterBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremoteharvesterbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":158,"kind":2048,"name":"generateRemoteHarvesterOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremoteharvesteroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":159,"kind":2048,"name":"generateRemoteReserverBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremotereserverbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":160,"kind":2048,"name":"generateRemoteReserverOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremotereserveroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":161,"kind":2048,"name":"generateRemoteColonizerBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremotecolonizerbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":162,"kind":2048,"name":"generateRemoteColonizerOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremotecolonizeroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":163,"kind":2048,"name":"generateRemoteDefenderBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremotedefenderbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":164,"kind":2048,"name":"generateRemoteDefenderOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremotedefenderoptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":165,"kind":2048,"name":"generateZealotBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatezealotbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":166,"kind":2048,"name":"generateZealotOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatezealotoptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":167,"kind":2048,"name":"generateMedicBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatemedicbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":168,"kind":2048,"name":"generateMedicOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatemedicoptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":169,"kind":2048,"name":"generateStalkerBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatestalkerbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":170,"kind":2048,"name":"generateStalkerOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatestalkeroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":171,"kind":2048,"name":"getDefaultCreepOptionsCiv","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#getdefaultcreepoptionsciv","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":172,"kind":2048,"name":"getDefaultCreepOptionsMili","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#getdefaultcreepoptionsmili","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":173,"kind":2048,"name":"generateDefaultCreepMemory","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatedefaultcreepmemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":174,"kind":1,"name":"\"src/Api/Spawn.Api\"","url":"modules/_src_api_spawn_api_.html","classes":"tsd-kind-external-module"},{"id":175,"kind":128,"name":"SpawnApi","url":"classes/_src_api_spawn_api_.spawnapi.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Api/Spawn.Api\""},{"id":176,"kind":2048,"name":"getCreepCount","url":"classes/_src_api_spawn_api_.spawnapi.html#getcreepcount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":177,"kind":2048,"name":"getCreepLimits","url":"classes/_src_api_spawn_api_.spawnapi.html#getcreeplimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":178,"kind":2048,"name":"generateDomesticCreepLimits","url":"classes/_src_api_spawn_api_.spawnapi.html#generatedomesticcreeplimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":179,"kind":2048,"name":"generateRemoteCreepLimits","url":"classes/_src_api_spawn_api_.spawnapi.html#generateremotecreeplimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":180,"kind":2048,"name":"generateMilitaryCreepLimits","url":"classes/_src_api_spawn_api_.spawnapi.html#generatemilitarycreeplimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":181,"kind":2048,"name":"setCreepLimits","url":"classes/_src_api_spawn_api_.spawnapi.html#setcreeplimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":182,"kind":2048,"name":"getOpenSpawn","url":"classes/_src_api_spawn_api_.spawnapi.html#getopenspawn","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":183,"kind":2048,"name":"getNextCreep","url":"classes/_src_api_spawn_api_.spawnapi.html#getnextcreep","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":184,"kind":2048,"name":"spawnNextCreep","url":"classes/_src_api_spawn_api_.spawnapi.html#spawnnextcreep","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":185,"kind":2048,"name":"getEnergyCostOfBody","url":"classes/_src_api_spawn_api_.spawnapi.html#getenergycostofbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":186,"kind":2048,"name":"getTier","url":"classes/_src_api_spawn_api_.spawnapi.html#gettier","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":187,"kind":2048,"name":"generateCreepOptions","url":"classes/_src_api_spawn_api_.spawnapi.html#generatecreepoptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-private tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":188,"kind":2048,"name":"generateCreepBody","url":"classes/_src_api_spawn_api_.spawnapi.html#generatecreepbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":189,"kind":2048,"name":"getCreepBody","url":"classes/_src_api_spawn_api_.spawnapi.html#getcreepbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":190,"kind":2048,"name":"getActiveMiners","url":"classes/_src_api_spawn_api_.spawnapi.html#getactiveminers","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-private tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":191,"kind":2048,"name":"generateSquadUUID","url":"classes/_src_api_spawn_api_.spawnapi.html#generatesquaduuid","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":192,"kind":1,"name":"\"src/Helpers/CreepHelper\"","url":"modules/_src_helpers_creephelper_.html","classes":"tsd-kind-external-module"},{"id":193,"kind":128,"name":"CreepHelper","url":"classes/_src_helpers_creephelper_.creephelper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/CreepHelper\""},{"id":194,"kind":1,"name":"\"src/Helpers/StructureHelper\"","url":"modules/_src_helpers_structurehelper_.html","classes":"tsd-kind-external-module"},{"id":195,"kind":128,"name":"StructureHelper","url":"classes/_src_helpers_structurehelper_.structurehelper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/StructureHelper\""},{"id":196,"kind":1,"name":"\"src/utils/StringMapApi\"","url":"modules/_src_utils_stringmapapi_.html","classes":"tsd-kind-external-module"},{"id":197,"kind":1,"name":"\"test/integration/helper\"","url":"modules/_test_integration_helper_.html","classes":"tsd-kind-external-module"},{"id":198,"kind":128,"name":"IntegrationTestHelper","url":"classes/_test_integration_helper_.integrationtesthelper.html","classes":"tsd-kind-class tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"test/integration/helper\""},{"id":199,"kind":1024,"name":"_server","url":"classes/_test_integration_helper_.integrationtesthelper.html#_server","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-private tsd-is-not-exported","parent":"\"test/integration/helper\".IntegrationTestHelper"},{"id":200,"kind":1024,"name":"_player","url":"classes/_test_integration_helper_.integrationtesthelper.html#_player","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-private tsd-is-not-exported","parent":"\"test/integration/helper\".IntegrationTestHelper"},{"id":201,"kind":262144,"name":"server","url":"classes/_test_integration_helper_.integrationtesthelper.html#server","classes":"tsd-kind-get-signature tsd-parent-kind-class tsd-is-not-exported","parent":"\"test/integration/helper\".IntegrationTestHelper"},{"id":202,"kind":262144,"name":"player","url":"classes/_test_integration_helper_.integrationtesthelper.html#player","classes":"tsd-kind-get-signature tsd-parent-kind-class tsd-is-not-exported","parent":"\"test/integration/helper\".IntegrationTestHelper"},{"id":203,"kind":2048,"name":"beforeEach","url":"classes/_test_integration_helper_.integrationtesthelper.html#beforeeach","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-not-exported","parent":"\"test/integration/helper\".IntegrationTestHelper"},{"id":204,"kind":2048,"name":"afterEach","url":"classes/_test_integration_helper_.integrationtesthelper.html#aftereach","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-not-exported","parent":"\"test/integration/helper\".IntegrationTestHelper"},{"id":205,"kind":32,"name":"readFileSync","url":"modules/_test_integration_helper_.html#readfilesync","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"test/integration/helper\""},{"id":206,"kind":32,"name":"_","url":"modules/_test_integration_helper_.html#_","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"test/integration/helper\""},{"id":207,"kind":32,"name":"ScreepsServer","url":"modules/_test_integration_helper_.html#screepsserver","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"test/integration/helper\""},{"id":208,"kind":32,"name":"stdHooks","url":"modules/_test_integration_helper_.html#stdhooks","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"test/integration/helper\""},{"id":209,"kind":32,"name":"DIST_MAIN_JS","url":"modules/_test_integration_helper_.html#dist_main_js","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"test/integration/helper\""},{"id":210,"kind":32,"name":"helper","url":"modules/_test_integration_helper_.html#helper","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/integration/helper\""},{"id":211,"kind":1,"name":"\"test/integration/SpawnApi.test\"","url":"modules/_test_integration_spawnapi_test_.html","classes":"tsd-kind-external-module"},{"id":212,"kind":1,"name":"\"test/integration/integration.test\"","url":"modules/_test_integration_integration_test_.html","classes":"tsd-kind-external-module"},{"id":213,"kind":1,"name":"\"test/unit/mock\"","url":"modules/_test_unit_mock_.html","classes":"tsd-kind-external-module"},{"id":214,"kind":2097152,"name":"Game","url":"modules/_test_unit_mock_.html#game","classes":"tsd-kind-object-literal tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":215,"kind":32,"name":"creeps","url":"modules/_test_unit_mock_.html#game.creeps","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".Game"},{"id":216,"kind":32,"name":"rooms","url":"modules/_test_unit_mock_.html#game.rooms","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".Game"},{"id":217,"kind":32,"name":"spawns","url":"modules/_test_unit_mock_.html#game.spawns","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".Game"},{"id":218,"kind":65536,"name":"__type","url":"modules/_test_unit_mock_.html#game.spawns.__type","classes":"tsd-kind-type-literal tsd-parent-kind-variable tsd-is-not-exported","parent":"\"test/unit/mock\".Game.spawns"},{"id":219,"kind":32,"name":"time","url":"modules/_test_unit_mock_.html#game.time","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".Game"},{"id":220,"kind":2097152,"name":"Memory","url":"modules/_test_unit_mock_.html#memory","classes":"tsd-kind-object-literal tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":221,"kind":32,"name":"creeps","url":"modules/_test_unit_mock_.html#memory.creeps-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".Memory"},{"id":222,"kind":32,"name":"OK","url":"modules/_test_unit_mock_.html#ok","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":223,"kind":32,"name":"ERR_NOT_OWNER","url":"modules/_test_unit_mock_.html#err_not_owner","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":224,"kind":32,"name":"ERR_NO_PATH","url":"modules/_test_unit_mock_.html#err_no_path","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":225,"kind":32,"name":"ERR_NAME_EXISTS","url":"modules/_test_unit_mock_.html#err_name_exists","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":226,"kind":32,"name":"ERR_BUSY","url":"modules/_test_unit_mock_.html#err_busy","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":227,"kind":32,"name":"ERR_NOT_FOUND","url":"modules/_test_unit_mock_.html#err_not_found","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":228,"kind":32,"name":"ERR_NOT_ENOUGH_RESOURCES","url":"modules/_test_unit_mock_.html#err_not_enough_resources","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":229,"kind":32,"name":"ERR_NOT_ENOUGH_ENERGY","url":"modules/_test_unit_mock_.html#err_not_enough_energy","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":230,"kind":32,"name":"ERR_INVALID_TARGET","url":"modules/_test_unit_mock_.html#err_invalid_target","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":231,"kind":32,"name":"ERR_FULL","url":"modules/_test_unit_mock_.html#err_full","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":232,"kind":32,"name":"ERR_NOT_IN_RANGE","url":"modules/_test_unit_mock_.html#err_not_in_range","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":233,"kind":32,"name":"ERR_INVALID_ARGS","url":"modules/_test_unit_mock_.html#err_invalid_args","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":234,"kind":32,"name":"ERR_TIRED","url":"modules/_test_unit_mock_.html#err_tired","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":235,"kind":32,"name":"ERR_NO_BODYPART","url":"modules/_test_unit_mock_.html#err_no_bodypart","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":236,"kind":32,"name":"ERR_NOT_ENOUGH_EXTENSIONS","url":"modules/_test_unit_mock_.html#err_not_enough_extensions","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":237,"kind":32,"name":"ERR_RCL_NOT_ENOUGH","url":"modules/_test_unit_mock_.html#err_rcl_not_enough","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":238,"kind":32,"name":"ERR_GCL_NOT_ENOUGH","url":"modules/_test_unit_mock_.html#err_gcl_not_enough","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":239,"kind":32,"name":"FIND_EXIT_TOP","url":"modules/_test_unit_mock_.html#find_exit_top","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":240,"kind":32,"name":"FIND_EXIT_RIGHT","url":"modules/_test_unit_mock_.html#find_exit_right","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":241,"kind":32,"name":"FIND_EXIT_BOTTOM","url":"modules/_test_unit_mock_.html#find_exit_bottom","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":242,"kind":32,"name":"FIND_EXIT_LEFT","url":"modules/_test_unit_mock_.html#find_exit_left","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":243,"kind":32,"name":"FIND_EXIT","url":"modules/_test_unit_mock_.html#find_exit","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":244,"kind":32,"name":"FIND_CREEPS","url":"modules/_test_unit_mock_.html#find_creeps","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":245,"kind":32,"name":"FIND_MY_CREEPS","url":"modules/_test_unit_mock_.html#find_my_creeps","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":246,"kind":32,"name":"FIND_HOSTILE_CREEPS","url":"modules/_test_unit_mock_.html#find_hostile_creeps","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":247,"kind":32,"name":"FIND_SOURCES_ACTIVE","url":"modules/_test_unit_mock_.html#find_sources_active","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":248,"kind":32,"name":"FIND_SOURCES","url":"modules/_test_unit_mock_.html#find_sources","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":249,"kind":32,"name":"FIND_DROPPED_ENERGY","url":"modules/_test_unit_mock_.html#find_dropped_energy","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":250,"kind":32,"name":"FIND_DROPPED_RESOURCES","url":"modules/_test_unit_mock_.html#find_dropped_resources","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":251,"kind":32,"name":"FIND_STRUCTURES","url":"modules/_test_unit_mock_.html#find_structures","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":252,"kind":32,"name":"FIND_MY_STRUCTURES","url":"modules/_test_unit_mock_.html#find_my_structures","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":253,"kind":32,"name":"FIND_HOSTILE_STRUCTURES","url":"modules/_test_unit_mock_.html#find_hostile_structures","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":254,"kind":32,"name":"FIND_FLAGS","url":"modules/_test_unit_mock_.html#find_flags","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":255,"kind":32,"name":"FIND_CONSTRUCTION_SITES","url":"modules/_test_unit_mock_.html#find_construction_sites","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":256,"kind":32,"name":"FIND_MY_SPAWNS","url":"modules/_test_unit_mock_.html#find_my_spawns","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":257,"kind":32,"name":"FIND_HOSTILE_SPAWNS","url":"modules/_test_unit_mock_.html#find_hostile_spawns","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":258,"kind":32,"name":"FIND_MY_CONSTRUCTION_SITES","url":"modules/_test_unit_mock_.html#find_my_construction_sites","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":259,"kind":32,"name":"FIND_HOSTILE_CONSTRUCTION_SITES","url":"modules/_test_unit_mock_.html#find_hostile_construction_sites","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":260,"kind":32,"name":"FIND_MINERALS","url":"modules/_test_unit_mock_.html#find_minerals","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":261,"kind":32,"name":"FIND_NUKES","url":"modules/_test_unit_mock_.html#find_nukes","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":262,"kind":32,"name":"FIND_TOMBSTONES","url":"modules/_test_unit_mock_.html#find_tombstones","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":263,"kind":32,"name":"TOP","url":"modules/_test_unit_mock_.html#top","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":264,"kind":32,"name":"TOP_RIGHT","url":"modules/_test_unit_mock_.html#top_right","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":265,"kind":32,"name":"RIGHT","url":"modules/_test_unit_mock_.html#right","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":266,"kind":32,"name":"BOTTOM_RIGHT","url":"modules/_test_unit_mock_.html#bottom_right","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":267,"kind":32,"name":"BOTTOM","url":"modules/_test_unit_mock_.html#bottom","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":268,"kind":32,"name":"BOTTOM_LEFT","url":"modules/_test_unit_mock_.html#bottom_left","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":269,"kind":32,"name":"LEFT","url":"modules/_test_unit_mock_.html#left","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":270,"kind":32,"name":"TOP_LEFT","url":"modules/_test_unit_mock_.html#top_left","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":271,"kind":32,"name":"COLOR_RED","url":"modules/_test_unit_mock_.html#color_red","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":272,"kind":32,"name":"COLOR_PURPLE","url":"modules/_test_unit_mock_.html#color_purple","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":273,"kind":32,"name":"COLOR_BLUE","url":"modules/_test_unit_mock_.html#color_blue","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":274,"kind":32,"name":"COLOR_CYAN","url":"modules/_test_unit_mock_.html#color_cyan","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":275,"kind":32,"name":"COLOR_GREEN","url":"modules/_test_unit_mock_.html#color_green","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":276,"kind":32,"name":"COLOR_YELLOW","url":"modules/_test_unit_mock_.html#color_yellow","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":277,"kind":32,"name":"COLOR_ORANGE","url":"modules/_test_unit_mock_.html#color_orange","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":278,"kind":32,"name":"COLOR_BROWN","url":"modules/_test_unit_mock_.html#color_brown","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":279,"kind":32,"name":"COLOR_GREY","url":"modules/_test_unit_mock_.html#color_grey","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":280,"kind":32,"name":"COLOR_WHITE","url":"modules/_test_unit_mock_.html#color_white","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":281,"kind":32,"name":"CREEP_SPAWN_TIME","url":"modules/_test_unit_mock_.html#creep_spawn_time","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":282,"kind":32,"name":"CREEP_LIFE_TIME","url":"modules/_test_unit_mock_.html#creep_life_time","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":283,"kind":32,"name":"CREEP_CLAIM_LIFE_TIME","url":"modules/_test_unit_mock_.html#creep_claim_life_time","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":284,"kind":32,"name":"CREEP_CORPSE_RATE","url":"modules/_test_unit_mock_.html#creep_corpse_rate","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":285,"kind":32,"name":"OBSTACLE_OBJECT_TYPES","url":"modules/_test_unit_mock_.html#obstacle_object_types","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":286,"kind":32,"name":"ENERGY_REGEN_TIME","url":"modules/_test_unit_mock_.html#energy_regen_time","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":287,"kind":32,"name":"ENERGY_DECAY","url":"modules/_test_unit_mock_.html#energy_decay","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":288,"kind":32,"name":"REPAIR_COST","url":"modules/_test_unit_mock_.html#repair_cost","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":289,"kind":32,"name":"RAMPART_DECAY_AMOUNT","url":"modules/_test_unit_mock_.html#rampart_decay_amount","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":290,"kind":32,"name":"RAMPART_DECAY_TIME","url":"modules/_test_unit_mock_.html#rampart_decay_time","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":291,"kind":32,"name":"RAMPART_HITS","url":"modules/_test_unit_mock_.html#rampart_hits","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":292,"kind":2097152,"name":"RAMPART_HITS_MAX","url":"modules/_test_unit_mock_.html#rampart_hits_max","classes":"tsd-kind-object-literal tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":293,"kind":32,"name":"2","url":"modules/_test_unit_mock_.html#rampart_hits_max.2-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".RAMPART_HITS_MAX"},{"id":294,"kind":32,"name":"3","url":"modules/_test_unit_mock_.html#rampart_hits_max.3-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".RAMPART_HITS_MAX"},{"id":295,"kind":32,"name":"4","url":"modules/_test_unit_mock_.html#rampart_hits_max.4-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".RAMPART_HITS_MAX"},{"id":296,"kind":32,"name":"5","url":"modules/_test_unit_mock_.html#rampart_hits_max.5-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".RAMPART_HITS_MAX"},{"id":297,"kind":32,"name":"6","url":"modules/_test_unit_mock_.html#rampart_hits_max.6-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".RAMPART_HITS_MAX"},{"id":298,"kind":32,"name":"7","url":"modules/_test_unit_mock_.html#rampart_hits_max.7-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".RAMPART_HITS_MAX"},{"id":299,"kind":32,"name":"8","url":"modules/_test_unit_mock_.html#rampart_hits_max.8-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".RAMPART_HITS_MAX"},{"id":300,"kind":32,"name":"SPAWN_HITS","url":"modules/_test_unit_mock_.html#spawn_hits","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":301,"kind":32,"name":"SPAWN_ENERGY_START","url":"modules/_test_unit_mock_.html#spawn_energy_start","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":302,"kind":32,"name":"SPAWN_ENERGY_CAPACITY","url":"modules/_test_unit_mock_.html#spawn_energy_capacity","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":303,"kind":32,"name":"SOURCE_ENERGY_CAPACITY","url":"modules/_test_unit_mock_.html#source_energy_capacity","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":304,"kind":32,"name":"SOURCE_ENERGY_NEUTRAL_CAPACITY","url":"modules/_test_unit_mock_.html#source_energy_neutral_capacity","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":305,"kind":32,"name":"SOURCE_ENERGY_KEEPER_CAPACITY","url":"modules/_test_unit_mock_.html#source_energy_keeper_capacity","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":306,"kind":32,"name":"WALL_HITS","url":"modules/_test_unit_mock_.html#wall_hits","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":307,"kind":32,"name":"WALL_HITS_MAX","url":"modules/_test_unit_mock_.html#wall_hits_max","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":308,"kind":32,"name":"EXTENSION_HITS","url":"modules/_test_unit_mock_.html#extension_hits","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":309,"kind":2097152,"name":"EXTENSION_ENERGY_CAPACITY","url":"modules/_test_unit_mock_.html#extension_energy_capacity","classes":"tsd-kind-object-literal tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":310,"kind":32,"name":"0","url":"modules/_test_unit_mock_.html#extension_energy_capacity.0","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":311,"kind":32,"name":"1","url":"modules/_test_unit_mock_.html#extension_energy_capacity.1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":312,"kind":32,"name":"2","url":"modules/_test_unit_mock_.html#extension_energy_capacity.2","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":313,"kind":32,"name":"3","url":"modules/_test_unit_mock_.html#extension_energy_capacity.3","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":314,"kind":32,"name":"4","url":"modules/_test_unit_mock_.html#extension_energy_capacity.4","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":315,"kind":32,"name":"5","url":"modules/_test_unit_mock_.html#extension_energy_capacity.5","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":316,"kind":32,"name":"6","url":"modules/_test_unit_mock_.html#extension_energy_capacity.6","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":317,"kind":32,"name":"7","url":"modules/_test_unit_mock_.html#extension_energy_capacity.7","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":318,"kind":32,"name":"8","url":"modules/_test_unit_mock_.html#extension_energy_capacity.8","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":319,"kind":32,"name":"ROAD_HITS","url":"modules/_test_unit_mock_.html#road_hits","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":320,"kind":32,"name":"ROAD_WEAROUT","url":"modules/_test_unit_mock_.html#road_wearout","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":321,"kind":32,"name":"ROAD_DECAY_AMOUNT","url":"modules/_test_unit_mock_.html#road_decay_amount","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":322,"kind":32,"name":"ROAD_DECAY_TIME","url":"modules/_test_unit_mock_.html#road_decay_time","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":323,"kind":32,"name":"LINK_HITS","url":"modules/_test_unit_mock_.html#link_hits","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":324,"kind":32,"name":"LINK_HITS_MAX","url":"modules/_test_unit_mock_.html#link_hits_max","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":325,"kind":32,"name":"LINK_CAPACITY","url":"modules/_test_unit_mock_.html#link_capacity","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":326,"kind":32,"name":"LINK_COOLDOWN","url":"modules/_test_unit_mock_.html#link_cooldown","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":327,"kind":32,"name":"LINK_LOSS_RATIO","url":"modules/_test_unit_mock_.html#link_loss_ratio","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":328,"kind":32,"name":"STORAGE_CAPACITY","url":"modules/_test_unit_mock_.html#storage_capacity","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":329,"kind":32,"name":"STORAGE_HITS","url":"modules/_test_unit_mock_.html#storage_hits","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":330,"kind":32,"name":"CARRY_CAPACITY","url":"modules/_test_unit_mock_.html#carry_capacity","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":331,"kind":32,"name":"HARVEST_POWER","url":"modules/_test_unit_mock_.html#harvest_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":332,"kind":32,"name":"HARVEST_MINERAL_POWER","url":"modules/_test_unit_mock_.html#harvest_mineral_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":333,"kind":32,"name":"REPAIR_POWER","url":"modules/_test_unit_mock_.html#repair_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":334,"kind":32,"name":"DISMANTLE_POWER","url":"modules/_test_unit_mock_.html#dismantle_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":335,"kind":32,"name":"BUILD_POWER","url":"modules/_test_unit_mock_.html#build_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":336,"kind":32,"name":"ATTACK_POWER","url":"modules/_test_unit_mock_.html#attack_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":337,"kind":32,"name":"UPGRADE_CONTROLLER_POWER","url":"modules/_test_unit_mock_.html#upgrade_controller_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":338,"kind":32,"name":"RANGED_ATTACK_POWER","url":"modules/_test_unit_mock_.html#ranged_attack_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":339,"kind":32,"name":"HEAL_POWER","url":"modules/_test_unit_mock_.html#heal_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":340,"kind":32,"name":"RANGED_HEAL_POWER","url":"modules/_test_unit_mock_.html#ranged_heal_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":341,"kind":32,"name":"DISMANTLE_COST","url":"modules/_test_unit_mock_.html#dismantle_cost","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":342,"kind":32,"name":"MOVE","url":"modules/_test_unit_mock_.html#move","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":343,"kind":32,"name":"WORK","url":"modules/_test_unit_mock_.html#work","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":344,"kind":32,"name":"CARRY","url":"modules/_test_unit_mock_.html#carry","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":345,"kind":32,"name":"ATTACK","url":"modules/_test_unit_mock_.html#attack","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":346,"kind":32,"name":"RANGED_ATTACK","url":"modules/_test_unit_mock_.html#ranged_attack","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":347,"kind":32,"name":"TOUGH","url":"modules/_test_unit_mock_.html#tough","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":348,"kind":32,"name":"HEAL","url":"modules/_test_unit_mock_.html#heal","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":349,"kind":32,"name":"CLAIM","url":"modules/_test_unit_mock_.html#claim","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":350,"kind":32,"name":"BODYPARTS_ALL","url":"modules/_test_unit_mock_.html#bodyparts_all","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":351,"kind":32,"name":"CONSTRUCTION_COST_ROAD_SWAMP_RATIO","url":"modules/_test_unit_mock_.html#construction_cost_road_swamp_ratio","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":352,"kind":32,"name":"STRUCTURE_EXTENSION","url":"modules/_test_unit_mock_.html#structure_extension","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":353,"kind":32,"name":"STRUCTURE_RAMPART","url":"modules/_test_unit_mock_.html#structure_rampart","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":354,"kind":32,"name":"STRUCTURE_ROAD","url":"modules/_test_unit_mock_.html#structure_road","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":355,"kind":32,"name":"STRUCTURE_SPAWN","url":"modules/_test_unit_mock_.html#structure_spawn","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":356,"kind":32,"name":"STRUCTURE_LINK","url":"modules/_test_unit_mock_.html#structure_link","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":357,"kind":32,"name":"STRUCTURE_WALL","url":"modules/_test_unit_mock_.html#structure_wall","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":358,"kind":32,"name":"STRUCTURE_KEEPER_LAIR","url":"modules/_test_unit_mock_.html#structure_keeper_lair","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":359,"kind":32,"name":"STRUCTURE_CONTROLLER","url":"modules/_test_unit_mock_.html#structure_controller","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":360,"kind":32,"name":"STRUCTURE_STORAGE","url":"modules/_test_unit_mock_.html#structure_storage","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":361,"kind":32,"name":"STRUCTURE_TOWER","url":"modules/_test_unit_mock_.html#structure_tower","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":362,"kind":32,"name":"STRUCTURE_OBSERVER","url":"modules/_test_unit_mock_.html#structure_observer","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":363,"kind":32,"name":"STRUCTURE_POWER_BANK","url":"modules/_test_unit_mock_.html#structure_power_bank","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":364,"kind":32,"name":"STRUCTURE_POWER_SPAWN","url":"modules/_test_unit_mock_.html#structure_power_spawn","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":365,"kind":32,"name":"STRUCTURE_EXTRACTOR","url":"modules/_test_unit_mock_.html#structure_extractor","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":366,"kind":32,"name":"STRUCTURE_LAB","url":"modules/_test_unit_mock_.html#structure_lab","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":367,"kind":32,"name":"STRUCTURE_TERMINAL","url":"modules/_test_unit_mock_.html#structure_terminal","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":368,"kind":32,"name":"STRUCTURE_CONTAINER","url":"modules/_test_unit_mock_.html#structure_container","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":369,"kind":32,"name":"STRUCTURE_NUKER","url":"modules/_test_unit_mock_.html#structure_nuker","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":370,"kind":32,"name":"STRUCTURE_PORTAL","url":"modules/_test_unit_mock_.html#structure_portal","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":371,"kind":32,"name":"RESOURCE_ENERGY","url":"modules/_test_unit_mock_.html#resource_energy","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":372,"kind":32,"name":"RESOURCE_POWER","url":"modules/_test_unit_mock_.html#resource_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":373,"kind":32,"name":"RESOURCE_UTRIUM","url":"modules/_test_unit_mock_.html#resource_utrium","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":374,"kind":32,"name":"RESOURCE_LEMERGIUM","url":"modules/_test_unit_mock_.html#resource_lemergium","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":375,"kind":32,"name":"RESOURCE_KEANIUM","url":"modules/_test_unit_mock_.html#resource_keanium","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":376,"kind":32,"name":"RESOURCE_GHODIUM","url":"modules/_test_unit_mock_.html#resource_ghodium","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":377,"kind":32,"name":"RESOURCE_ZYNTHIUM","url":"modules/_test_unit_mock_.html#resource_zynthium","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":378,"kind":32,"name":"RESOURCE_OXYGEN","url":"modules/_test_unit_mock_.html#resource_oxygen","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":379,"kind":32,"name":"RESOURCE_HYDROGEN","url":"modules/_test_unit_mock_.html#resource_hydrogen","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":380,"kind":32,"name":"RESOURCE_CATALYST","url":"modules/_test_unit_mock_.html#resource_catalyst","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":381,"kind":32,"name":"RESOURCE_HYDROXIDE","url":"modules/_test_unit_mock_.html#resource_hydroxide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":382,"kind":32,"name":"RESOURCE_ZYNTHIUM_KEANITE","url":"modules/_test_unit_mock_.html#resource_zynthium_keanite","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":383,"kind":32,"name":"RESOURCE_UTRIUM_LEMERGITE","url":"modules/_test_unit_mock_.html#resource_utrium_lemergite","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":384,"kind":32,"name":"RESOURCE_UTRIUM_HYDRIDE","url":"modules/_test_unit_mock_.html#resource_utrium_hydride","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":385,"kind":32,"name":"RESOURCE_UTRIUM_OXIDE","url":"modules/_test_unit_mock_.html#resource_utrium_oxide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":386,"kind":32,"name":"RESOURCE_KEANIUM_HYDRIDE","url":"modules/_test_unit_mock_.html#resource_keanium_hydride","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":387,"kind":32,"name":"RESOURCE_KEANIUM_OXIDE","url":"modules/_test_unit_mock_.html#resource_keanium_oxide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":388,"kind":32,"name":"RESOURCE_LEMERGIUM_HYDRIDE","url":"modules/_test_unit_mock_.html#resource_lemergium_hydride","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":389,"kind":32,"name":"RESOURCE_LEMERGIUM_OXIDE","url":"modules/_test_unit_mock_.html#resource_lemergium_oxide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":390,"kind":32,"name":"RESOURCE_ZYNTHIUM_HYDRIDE","url":"modules/_test_unit_mock_.html#resource_zynthium_hydride","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":391,"kind":32,"name":"RESOURCE_ZYNTHIUM_OXIDE","url":"modules/_test_unit_mock_.html#resource_zynthium_oxide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":392,"kind":32,"name":"RESOURCE_GHODIUM_HYDRIDE","url":"modules/_test_unit_mock_.html#resource_ghodium_hydride","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":393,"kind":32,"name":"RESOURCE_GHODIUM_OXIDE","url":"modules/_test_unit_mock_.html#resource_ghodium_oxide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":394,"kind":32,"name":"RESOURCE_UTRIUM_ACID","url":"modules/_test_unit_mock_.html#resource_utrium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":395,"kind":32,"name":"RESOURCE_UTRIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_utrium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":396,"kind":32,"name":"RESOURCE_KEANIUM_ACID","url":"modules/_test_unit_mock_.html#resource_keanium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":397,"kind":32,"name":"RESOURCE_KEANIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_keanium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":398,"kind":32,"name":"RESOURCE_LEMERGIUM_ACID","url":"modules/_test_unit_mock_.html#resource_lemergium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":399,"kind":32,"name":"RESOURCE_LEMERGIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_lemergium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":400,"kind":32,"name":"RESOURCE_ZYNTHIUM_ACID","url":"modules/_test_unit_mock_.html#resource_zynthium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":401,"kind":32,"name":"RESOURCE_ZYNTHIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_zynthium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":402,"kind":32,"name":"RESOURCE_GHODIUM_ACID","url":"modules/_test_unit_mock_.html#resource_ghodium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":403,"kind":32,"name":"RESOURCE_GHODIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_ghodium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":404,"kind":32,"name":"RESOURCE_CATALYZED_UTRIUM_ACID","url":"modules/_test_unit_mock_.html#resource_catalyzed_utrium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":405,"kind":32,"name":"RESOURCE_CATALYZED_UTRIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_catalyzed_utrium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":406,"kind":32,"name":"RESOURCE_CATALYZED_KEANIUM_ACID","url":"modules/_test_unit_mock_.html#resource_catalyzed_keanium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":407,"kind":32,"name":"RESOURCE_CATALYZED_KEANIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_catalyzed_keanium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":408,"kind":32,"name":"RESOURCE_CATALYZED_LEMERGIUM_ACID","url":"modules/_test_unit_mock_.html#resource_catalyzed_lemergium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":409,"kind":32,"name":"RESOURCE_CATALYZED_LEMERGIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_catalyzed_lemergium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":410,"kind":32,"name":"RESOURCE_CATALYZED_ZYNTHIUM_ACID","url":"modules/_test_unit_mock_.html#resource_catalyzed_zynthium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":411,"kind":32,"name":"RESOURCE_CATALYZED_ZYNTHIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_catalyzed_zynthium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":412,"kind":32,"name":"RESOURCE_CATALYZED_GHODIUM_ACID","url":"modules/_test_unit_mock_.html#resource_catalyzed_ghodium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":413,"kind":32,"name":"RESOURCE_CATALYZED_GHODIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_catalyzed_ghodium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":414,"kind":32,"name":"SUBSCRIPTION_TOKEN","url":"modules/_test_unit_mock_.html#subscription_token","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":415,"kind":32,"name":"SAFE_MODE_DURATION","url":"modules/_test_unit_mock_.html#safe_mode_duration","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":416,"kind":32,"name":"SAFE_MODE_COOLDOWN","url":"modules/_test_unit_mock_.html#safe_mode_cooldown","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":417,"kind":32,"name":"SAFE_MODE_COST","url":"modules/_test_unit_mock_.html#safe_mode_cost","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":418,"kind":32,"name":"LOOK_CREEPS","url":"modules/_test_unit_mock_.html#look_creeps","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":419,"kind":32,"name":"LOOK_ENERGY","url":"modules/_test_unit_mock_.html#look_energy","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":420,"kind":32,"name":"LOOK_RESOURCES","url":"modules/_test_unit_mock_.html#look_resources","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":421,"kind":32,"name":"LOOK_SOURCES","url":"modules/_test_unit_mock_.html#look_sources","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":422,"kind":32,"name":"LOOK_MINERALS","url":"modules/_test_unit_mock_.html#look_minerals","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":423,"kind":32,"name":"LOOK_STRUCTURES","url":"modules/_test_unit_mock_.html#look_structures","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":424,"kind":32,"name":"LOOK_FLAGS","url":"modules/_test_unit_mock_.html#look_flags","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":425,"kind":32,"name":"LOOK_CONSTRUCTION_SITES","url":"modules/_test_unit_mock_.html#look_construction_sites","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":426,"kind":32,"name":"LOOK_NUKES","url":"modules/_test_unit_mock_.html#look_nukes","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":427,"kind":32,"name":"LOOK_TERRAIN","url":"modules/_test_unit_mock_.html#look_terrain","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":428,"kind":32,"name":"LOOK_TOMBSTONES","url":"modules/_test_unit_mock_.html#look_tombstones","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":429,"kind":32,"name":"ORDER_SELL","url":"modules/_test_unit_mock_.html#order_sell","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":430,"kind":32,"name":"ORDER_BUY","url":"modules/_test_unit_mock_.html#order_buy","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":431,"kind":32,"name":"TOMBSTONE_DECAY_PER_PART","url":"modules/_test_unit_mock_.html#tombstone_decay_per_part","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":432,"kind":32,"name":"EVENT_ATTACK","url":"modules/_test_unit_mock_.html#event_attack","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":433,"kind":32,"name":"EVENT_OBJECT_DESTROYED","url":"modules/_test_unit_mock_.html#event_object_destroyed","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":434,"kind":32,"name":"EVENT_ATTACK_CONTROLLER","url":"modules/_test_unit_mock_.html#event_attack_controller","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":435,"kind":32,"name":"EVENT_BUILD","url":"modules/_test_unit_mock_.html#event_build","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":436,"kind":32,"name":"EVENT_HARVEST","url":"modules/_test_unit_mock_.html#event_harvest","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":437,"kind":32,"name":"EVENT_HEAL","url":"modules/_test_unit_mock_.html#event_heal","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":438,"kind":32,"name":"EVENT_REPAIR","url":"modules/_test_unit_mock_.html#event_repair","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":439,"kind":32,"name":"EVENT_RESERVE_CONTROLLER","url":"modules/_test_unit_mock_.html#event_reserve_controller","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":440,"kind":32,"name":"EVENT_UPGRADE_CONTROLLER","url":"modules/_test_unit_mock_.html#event_upgrade_controller","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":441,"kind":32,"name":"EVENT_EXIT","url":"modules/_test_unit_mock_.html#event_exit","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":442,"kind":32,"name":"EVENT_ATTACK_TYPE_MELEE","url":"modules/_test_unit_mock_.html#event_attack_type_melee","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":443,"kind":32,"name":"EVENT_ATTACK_TYPE_RANGED","url":"modules/_test_unit_mock_.html#event_attack_type_ranged","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":444,"kind":32,"name":"EVENT_ATTACK_TYPE_RANGED_MASS","url":"modules/_test_unit_mock_.html#event_attack_type_ranged_mass","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":445,"kind":32,"name":"EVENT_ATTACK_TYPE_DISMANTLE","url":"modules/_test_unit_mock_.html#event_attack_type_dismantle","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":446,"kind":32,"name":"EVENT_ATTACK_TYPE_HIT_BACK","url":"modules/_test_unit_mock_.html#event_attack_type_hit_back","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":447,"kind":32,"name":"EVENT_ATTACK_TYPE_NUKE","url":"modules/_test_unit_mock_.html#event_attack_type_nuke","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":448,"kind":32,"name":"EVENT_HEAL_TYPE_MELEE","url":"modules/_test_unit_mock_.html#event_heal_type_melee","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":449,"kind":32,"name":"EVENT_HEAL_TYPE_RANGED","url":"modules/_test_unit_mock_.html#event_heal_type_ranged","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":450,"kind":1,"name":"\"test/unit/SpawnApi.test\"","url":"modules/_test_unit_spawnapi_test_.html","classes":"tsd-kind-external-module"},{"id":451,"kind":1,"name":"\"test/unit/main.test\"","url":"modules/_test_unit_main_test_.html","classes":"tsd-kind-external-module"}]};
+            typedoc.search.data = {"kinds":{"1":"External module","32":"Variable","64":"Function","128":"Class","1024":"Property","2048":"Method","65536":"Type literal","262144":"Accessor","2097152":"Object literal"},"rows":[{"id":0,"kind":1,"name":"\"src/utils/config\"","url":"modules/_src_utils_config_.html","classes":"tsd-kind-external-module"},{"id":1,"kind":32,"name":"NO_CACHING_MEMORY","url":"modules/_src_utils_config_.html#no_caching_memory","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"src/utils/config\""},{"id":2,"kind":32,"name":"ALLOW_CUSTOM_ERRORS","url":"modules/_src_utils_config_.html#allow_custom_errors","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"src/utils/config\""},{"id":3,"kind":1,"name":"\"src/Helpers/UtilHelper\"","url":"modules/_src_helpers_utilhelper_.html","classes":"tsd-kind-external-module"},{"id":4,"kind":128,"name":"UtilHelper","url":"classes/_src_helpers_utilhelper_.utilhelper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/UtilHelper\""},{"id":5,"kind":2048,"name":"throwError","url":"classes/_src_helpers_utilhelper_.utilhelper.html#throwerror","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/UtilHelper\".UtilHelper"},{"id":6,"kind":1,"name":"\"src/Helpers/Normalize\"","url":"modules/_src_helpers_normalize_.html","classes":"tsd-kind-external-module"},{"id":7,"kind":128,"name":"Normalize","url":"classes/_src_helpers_normalize_.normalize.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/Normalize\""},{"id":8,"kind":2048,"name":"roomName","url":"classes/_src_helpers_normalize_.normalize.html#roomname","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/Normalize\".Normalize"},{"id":9,"kind":2048,"name":"roomObject","url":"classes/_src_helpers_normalize_.normalize.html#roomobject","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/Normalize\".Normalize"},{"id":10,"kind":2048,"name":"roomPos","url":"classes/_src_helpers_normalize_.normalize.html#roompos","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/Normalize\".Normalize"},{"id":11,"kind":2048,"name":"creepObject","url":"classes/_src_helpers_normalize_.normalize.html#creepobject","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/Normalize\".Normalize"},{"id":12,"kind":1,"name":"\"src/Helpers/MemoryHelper_Room\"","url":"modules/_src_helpers_memoryhelper_room_.html","classes":"tsd-kind-external-module"},{"id":13,"kind":128,"name":"MemoryHelper_Room","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/MemoryHelper_Room\""},{"id":14,"kind":2048,"name":"updateRoomMemory","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updateroommemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":15,"kind":2048,"name":"updateDependentRooms","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatedependentrooms","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":16,"kind":2048,"name":"updateHostileCreeps","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatehostilecreeps","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":17,"kind":2048,"name":"updateMyCreeps","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatemycreeps","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":18,"kind":2048,"name":"updateConstructionSites","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updateconstructionsites","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":19,"kind":2048,"name":"updateStructures","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatestructures","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":20,"kind":2048,"name":"updateSources","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatesources","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":21,"kind":2048,"name":"updateRoomState","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updateroomstate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":22,"kind":2048,"name":"updateDefcon","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatedefcon","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":23,"kind":2048,"name":"updateDomesticLimits","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatedomesticlimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":24,"kind":2048,"name":"updateRemoteLimits","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updateremotelimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":25,"kind":2048,"name":"updateMilitaryLimits","url":"classes/_src_helpers_memoryhelper_room_.memoryhelper_room.html#updatemilitarylimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper_Room\".MemoryHelper_Room"},{"id":26,"kind":1,"name":"\"src/Api/Memory.Api\"","url":"modules/_src_api_memory_api_.html","classes":"tsd-kind-external-module"},{"id":27,"kind":128,"name":"MemoryApi","url":"classes/_src_api_memory_api_.memoryapi.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Api/Memory.Api\""},{"id":28,"kind":2048,"name":"garbageCollection","url":"classes/_src_api_memory_api_.memoryapi.html#garbagecollection","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":29,"kind":2048,"name":"initRoomMemory","url":"classes/_src_api_memory_api_.memoryapi.html#initroommemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":30,"kind":2048,"name":"getRoomMemory","url":"classes/_src_api_memory_api_.memoryapi.html#getroommemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":31,"kind":2048,"name":"initCreepMemory","url":"classes/_src_api_memory_api_.memoryapi.html#initcreepmemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":32,"kind":2048,"name":"getMyCreeps","url":"classes/_src_api_memory_api_.memoryapi.html#getmycreeps","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":33,"kind":2048,"name":"getHostileCreeps","url":"classes/_src_api_memory_api_.memoryapi.html#gethostilecreeps","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":34,"kind":2048,"name":"getStructures","url":"classes/_src_api_memory_api_.memoryapi.html#getstructures","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":35,"kind":2048,"name":"getStructureOfType","url":"classes/_src_api_memory_api_.memoryapi.html#getstructureoftype","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":36,"kind":2048,"name":"getConstructionSites","url":"classes/_src_api_memory_api_.memoryapi.html#getconstructionsites","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":37,"kind":2048,"name":"getSources","url":"classes/_src_api_memory_api_.memoryapi.html#getsources","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":38,"kind":2048,"name":"getRemoteRooms","url":"classes/_src_api_memory_api_.memoryapi.html#getremoterooms","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":39,"kind":2048,"name":"getClaimRooms","url":"classes/_src_api_memory_api_.memoryapi.html#getclaimrooms","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":40,"kind":2048,"name":"getAttackRooms","url":"classes/_src_api_memory_api_.memoryapi.html#getattackrooms","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Memory.Api\".MemoryApi"},{"id":41,"kind":1,"name":"\"src/Helpers/RoomHelper\"","url":"modules/_src_helpers_roomhelper_.html","classes":"tsd-kind-external-module"},{"id":42,"kind":128,"name":"RoomHelper","url":"classes/_src_helpers_roomhelper_.roomhelper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/RoomHelper\""},{"id":43,"kind":2048,"name":"isOwnedRoom","url":"classes/_src_helpers_roomhelper_.roomhelper.html#isownedroom","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":44,"kind":2048,"name":"isAllyRoom","url":"classes/_src_helpers_roomhelper_.roomhelper.html#isallyroom","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":45,"kind":2048,"name":"isSourceKeeperRoom","url":"classes/_src_helpers_roomhelper_.roomhelper.html#issourcekeeperroom","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":46,"kind":2048,"name":"isHighwayRoom","url":"classes/_src_helpers_roomhelper_.roomhelper.html#ishighwayroom","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":47,"kind":2048,"name":"inTravelRange","url":"classes/_src_helpers_roomhelper_.roomhelper.html#intravelrange","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":48,"kind":2048,"name":"isExistInRoom","url":"classes/_src_helpers_roomhelper_.roomhelper.html#isexistinroom","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":49,"kind":2048,"name":"getStoredAmount","url":"classes/_src_helpers_roomhelper_.roomhelper.html#getstoredamount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":50,"kind":2048,"name":"getStoredCapacity","url":"classes/_src_helpers_roomhelper_.roomhelper.html#getstoredcapacity","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":51,"kind":2048,"name":"getTowerDamageAtRange","url":"classes/_src_helpers_roomhelper_.roomhelper.html#gettowerdamageatrange","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":52,"kind":2048,"name":"excecuteEveryTicks","url":"classes/_src_helpers_roomhelper_.roomhelper.html#excecuteeveryticks","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":53,"kind":2048,"name":"isContainerMining","url":"classes/_src_helpers_roomhelper_.roomhelper.html#iscontainermining","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":54,"kind":2048,"name":"isUpgraderLink","url":"classes/_src_helpers_roomhelper_.roomhelper.html#isupgraderlink","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":55,"kind":2048,"name":"chooseTowerTarget","url":"classes/_src_helpers_roomhelper_.roomhelper.html#choosetowertarget","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":56,"kind":2048,"name":"getWallLevelDifference","url":"classes/_src_helpers_roomhelper_.roomhelper.html#getwallleveldifference","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":57,"kind":2048,"name":"numHostileCreeps","url":"classes/_src_helpers_roomhelper_.roomhelper.html#numhostilecreeps","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":58,"kind":2048,"name":"numRemoteRooms","url":"classes/_src_helpers_roomhelper_.roomhelper.html#numremoterooms","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":59,"kind":2048,"name":"numClaimRooms","url":"classes/_src_helpers_roomhelper_.roomhelper.html#numclaimrooms","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":60,"kind":2048,"name":"numAttackRooms","url":"classes/_src_helpers_roomhelper_.roomhelper.html#numattackrooms","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":61,"kind":2048,"name":"numSources","url":"classes/_src_helpers_roomhelper_.roomhelper.html#numsources","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":62,"kind":2048,"name":"numRemoteSources","url":"classes/_src_helpers_roomhelper_.roomhelper.html#numremotesources","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":63,"kind":2048,"name":"numRemoteDefenders","url":"classes/_src_helpers_roomhelper_.roomhelper.html#numremotedefenders","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/RoomHelper\".RoomHelper"},{"id":64,"kind":1,"name":"\"src/Helpers/ConsoleCommands\"","url":"modules/_src_helpers_consolecommands_.html","classes":"tsd-kind-external-module"},{"id":65,"kind":2097152,"name":"ConsoleCommands","url":"modules/_src_helpers_consolecommands_.html#consolecommands","classes":"tsd-kind-object-literal tsd-parent-kind-external-module","parent":"\"src/Helpers/ConsoleCommands\""},{"id":66,"kind":64,"name":"removeConstructionSites","url":"modules/_src_helpers_consolecommands_.html#consolecommands.removeconstructionsites","classes":"tsd-kind-function tsd-parent-kind-object-literal","parent":"\"src/Helpers/ConsoleCommands\".ConsoleCommands"},{"id":67,"kind":64,"name":"removeFlags","url":"modules/_src_helpers_consolecommands_.html#consolecommands.removeflags","classes":"tsd-kind-function tsd-parent-kind-object-literal","parent":"\"src/Helpers/ConsoleCommands\".ConsoleCommands"},{"id":68,"kind":64,"name":"displayRoomStatus","url":"modules/_src_helpers_consolecommands_.html#consolecommands.displayroomstatus","classes":"tsd-kind-function tsd-parent-kind-object-literal","parent":"\"src/Helpers/ConsoleCommands\".ConsoleCommands"},{"id":69,"kind":64,"name":"killAllCreeps","url":"modules/_src_helpers_consolecommands_.html#consolecommands.killallcreeps","classes":"tsd-kind-function tsd-parent-kind-object-literal","parent":"\"src/Helpers/ConsoleCommands\".ConsoleCommands"},{"id":70,"kind":64,"name":"sendResource","url":"modules/_src_helpers_consolecommands_.html#consolecommands.sendresource","classes":"tsd-kind-function tsd-parent-kind-object-literal","parent":"\"src/Helpers/ConsoleCommands\".ConsoleCommands"},{"id":71,"kind":1,"name":"\"src/Managers/EmpireManager\"","url":"modules/_src_managers_empiremanager_.html","classes":"tsd-kind-external-module"},{"id":72,"kind":128,"name":"EmpireManager","url":"classes/_src_managers_empiremanager_.empiremanager.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Managers/EmpireManager\""},{"id":73,"kind":2048,"name":"runEmpireManager","url":"classes/_src_managers_empiremanager_.empiremanager.html#runempiremanager","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Managers/EmpireManager\".EmpireManager"},{"id":74,"kind":1,"name":"\"src/Managers/MemoryManagement\"","url":"modules/_src_managers_memorymanagement_.html","classes":"tsd-kind-external-module"},{"id":75,"kind":128,"name":"MemoryManager","url":"classes/_src_managers_memorymanagement_.memorymanager.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Managers/MemoryManagement\""},{"id":76,"kind":2048,"name":"runMemoryManager","url":"classes/_src_managers_memorymanagement_.memorymanager.html#runmemorymanager","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Managers/MemoryManagement\".MemoryManager"},{"id":77,"kind":1,"name":"\"src/Managers/RoomManager\"","url":"modules/_src_managers_roommanager_.html","classes":"tsd-kind-external-module"},{"id":78,"kind":128,"name":"RoomManager","url":"classes/_src_managers_roommanager_.roommanager.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Managers/RoomManager\""},{"id":79,"kind":2048,"name":"runRoomManager","url":"classes/_src_managers_roommanager_.roommanager.html#runroommanager","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Managers/RoomManager\".RoomManager"},{"id":80,"kind":1,"name":"\"src/Managers/SpawnManager\"","url":"modules/_src_managers_spawnmanager_.html","classes":"tsd-kind-external-module"},{"id":81,"kind":128,"name":"SpawnManager","url":"classes/_src_managers_spawnmanager_.spawnmanager.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Managers/SpawnManager\""},{"id":82,"kind":2048,"name":"runSpawnManager","url":"classes/_src_managers_spawnmanager_.spawnmanager.html#runspawnmanager","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Managers/SpawnManager\".SpawnManager"},{"id":83,"kind":1,"name":"\"src/utils/ErrorMapper\"","url":"modules/_src_utils_errormapper_.html","classes":"tsd-kind-external-module"},{"id":84,"kind":128,"name":"ErrorMapper","url":"classes/_src_utils_errormapper_.errormapper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/utils/ErrorMapper\""},{"id":85,"kind":1024,"name":"_consumer","url":"classes/_src_utils_errormapper_.errormapper.html#_consumer","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-private tsd-is-static","parent":"\"src/utils/ErrorMapper\".ErrorMapper"},{"id":86,"kind":262144,"name":"consumer","url":"classes/_src_utils_errormapper_.errormapper.html#consumer","classes":"tsd-kind-get-signature tsd-parent-kind-class tsd-is-static","parent":"\"src/utils/ErrorMapper\".ErrorMapper"},{"id":87,"kind":1024,"name":"cache","url":"classes/_src_utils_errormapper_.errormapper.html#cache","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-static","parent":"\"src/utils/ErrorMapper\".ErrorMapper"},{"id":88,"kind":65536,"name":"__type","url":"classes/_src_utils_errormapper_.errormapper.html#cache.__type","classes":"tsd-kind-type-literal tsd-parent-kind-property tsd-is-not-exported","parent":"\"src/utils/ErrorMapper\".ErrorMapper.cache"},{"id":89,"kind":2048,"name":"sourceMappedStackTrace","url":"classes/_src_utils_errormapper_.errormapper.html#sourcemappedstacktrace","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/utils/ErrorMapper\".ErrorMapper"},{"id":90,"kind":2048,"name":"wrapLoop","url":"classes/_src_utils_errormapper_.errormapper.html#wraploop","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/utils/ErrorMapper\".ErrorMapper"},{"id":91,"kind":1,"name":"\"src/main\"","url":"modules/_src_main_.html","classes":"tsd-kind-external-module"},{"id":92,"kind":32,"name":"loop","url":"modules/_src_main_.html#loop","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"src/main\""},{"id":93,"kind":65536,"name":"__type","url":"modules/_src_main_.html#loop.__type","classes":"tsd-kind-type-literal tsd-parent-kind-variable tsd-is-not-exported","parent":"\"src/main\".loop"},{"id":94,"kind":1,"name":"\"src/Api/CreepDomestic.Api\"","url":"modules/_src_api_creepdomestic_api_.html","classes":"tsd-kind-external-module"},{"id":95,"kind":128,"name":"CreepDomestic","url":"classes/_src_api_creepdomestic_api_.creepdomestic.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Api/CreepDomestic.Api\""},{"id":96,"kind":1,"name":"\"src/Api/CreepMili.Api\"","url":"modules/_src_api_creepmili_api_.html","classes":"tsd-kind-external-module"},{"id":97,"kind":128,"name":"CreepMili","url":"classes/_src_api_creepmili_api_.creepmili.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Api/CreepMili.Api\""},{"id":98,"kind":1,"name":"\"src/Api/CreepRemote.Api\"","url":"modules/_src_api_creepremote_api_.html","classes":"tsd-kind-external-module"},{"id":99,"kind":128,"name":"CreepRemote","url":"classes/_src_api_creepremote_api_.creepremote.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Api/CreepRemote.Api\""},{"id":100,"kind":1,"name":"\"src/Api/Empire.Api\"","url":"modules/_src_api_empire_api_.html","classes":"tsd-kind-external-module"},{"id":101,"kind":128,"name":"Empire","url":"classes/_src_api_empire_api_.empire.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Api/Empire.Api\""},{"id":102,"kind":2048,"name":"processNewFlags","url":"classes/_src_api_empire_api_.empire.html#processnewflags","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":103,"kind":2048,"name":"removeFlagMemory","url":"classes/_src_api_empire_api_.empire.html#removeflagmemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":104,"kind":2048,"name":"processNewRemoteFlag","url":"classes/_src_api_empire_api_.empire.html#processnewremoteflag","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":105,"kind":2048,"name":"processNewAttackFlag","url":"classes/_src_api_empire_api_.empire.html#processnewattackflag","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":106,"kind":2048,"name":"processNewClaimFlag","url":"classes/_src_api_empire_api_.empire.html#processnewclaimflag","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":107,"kind":2048,"name":"cleanDeadFlags","url":"classes/_src_api_empire_api_.empire.html#cleandeadflags","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":108,"kind":2048,"name":"findDependentRoom","url":"classes/_src_api_empire_api_.empire.html#finddependentroom","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":109,"kind":2048,"name":"findDependentRoomAuto","url":"classes/_src_api_empire_api_.empire.html#finddependentroomauto","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-private tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":110,"kind":2048,"name":"findDependentRoomManual","url":"classes/_src_api_empire_api_.empire.html#finddependentroommanual","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-private tsd-is-static","parent":"\"src/Api/Empire.Api\".Empire"},{"id":111,"kind":1,"name":"\"src/Helpers/MemoryHelper\"","url":"modules/_src_helpers_memoryhelper_.html","classes":"tsd-kind-external-module"},{"id":112,"kind":128,"name":"MemoryHelper","url":"classes/_src_helpers_memoryhelper_.memoryhelper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/MemoryHelper\""},{"id":113,"kind":2048,"name":"getStructureOfType","url":"classes/_src_helpers_memoryhelper_.memoryhelper.html#getstructureoftype","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper\".MemoryHelper"},{"id":114,"kind":2048,"name":"getCreepOfRole","url":"classes/_src_helpers_memoryhelper_.memoryhelper.html#getcreepofrole","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper\".MemoryHelper"},{"id":115,"kind":2048,"name":"clearCreepMemory","url":"classes/_src_helpers_memoryhelper_.memoryhelper.html#clearcreepmemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper\".MemoryHelper"},{"id":116,"kind":2048,"name":"clearRoomMemory","url":"classes/_src_helpers_memoryhelper_.memoryhelper.html#clearroommemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/MemoryHelper\".MemoryHelper"},{"id":117,"kind":1,"name":"\"src/Api/Room.Api\"","url":"modules/_src_api_room_api_.html","classes":"tsd-kind-external-module"},{"id":118,"kind":128,"name":"RoomApi","url":"classes/_src_api_room_api_.roomapi.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Api/Room.Api\""},{"id":119,"kind":2048,"name":"isHostilesInRoom","url":"classes/_src_api_room_api_.roomapi.html#ishostilesinroom","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":120,"kind":2048,"name":"setRoomState","url":"classes/_src_api_room_api_.roomapi.html#setroomstate","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":121,"kind":2048,"name":"runTowers","url":"classes/_src_api_room_api_.roomapi.html#runtowers","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":122,"kind":2048,"name":"setDefconLevel","url":"classes/_src_api_room_api_.roomapi.html#setdefconlevel","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":123,"kind":2048,"name":"getRepairTargets","url":"classes/_src_api_room_api_.roomapi.html#getrepairtargets","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":124,"kind":2048,"name":"getExtensionsNeedFilled","url":"classes/_src_api_room_api_.roomapi.html#getextensionsneedfilled","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":125,"kind":2048,"name":"getTowersNeedFilled","url":"classes/_src_api_room_api_.roomapi.html#gettowersneedfilled","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":126,"kind":2048,"name":"getWallRepairTargets","url":"classes/_src_api_room_api_.roomapi.html#getwallrepairtargets","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":127,"kind":2048,"name":"getOpenSources","url":"classes/_src_api_room_api_.roomapi.html#getopensources","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":128,"kind":2048,"name":"getMiningContainer","url":"classes/_src_api_room_api_.roomapi.html#getminingcontainer","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":129,"kind":2048,"name":"isFull","url":"classes/_src_api_room_api_.roomapi.html#isfull","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":130,"kind":2048,"name":"getWallHpLimit","url":"classes/_src_api_room_api_.roomapi.html#getwallhplimit","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-private tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":131,"kind":2048,"name":"createJobQueue","url":"classes/_src_api_room_api_.roomapi.html#createjobqueue","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":132,"kind":2048,"name":"createJobQueueWorker","url":"classes/_src_api_room_api_.roomapi.html#createjobqueueworker","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":133,"kind":2048,"name":"createJobQueueLorry","url":"classes/_src_api_room_api_.roomapi.html#createjobqueuelorry","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":134,"kind":2048,"name":"createEnergyQueue","url":"classes/_src_api_room_api_.roomapi.html#createenergyqueue","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":135,"kind":2048,"name":"runLinks","url":"classes/_src_api_room_api_.roomapi.html#runlinks","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":136,"kind":2048,"name":"runTerminal","url":"classes/_src_api_room_api_.roomapi.html#runterminal","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":137,"kind":2048,"name":"runLabs","url":"classes/_src_api_room_api_.roomapi.html#runlabs","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Room.Api\".RoomApi"},{"id":138,"kind":1,"name":"\"src/Helpers/SpawnHelper\"","url":"modules/_src_helpers_spawnhelper_.html","classes":"tsd-kind-external-module"},{"id":139,"kind":128,"name":"SpawnHelper","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/SpawnHelper\""},{"id":140,"kind":2048,"name":"verifyDescriptor","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#verifydescriptor","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":141,"kind":2048,"name":"generateParts","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateparts","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":142,"kind":2048,"name":"getBody_Grouped","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#getbody_grouped","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":143,"kind":2048,"name":"getBody_Collated","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#getbody_collated","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":144,"kind":2048,"name":"generateCreepName","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatecreepname","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":145,"kind":2048,"name":"generateMinerBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateminerbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":146,"kind":2048,"name":"generateMinerOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatemineroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":147,"kind":2048,"name":"generateHarvesterBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateharvesterbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":148,"kind":2048,"name":"generateHarvesterOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateharvesteroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":149,"kind":2048,"name":"generateWorkerBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateworkerbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":150,"kind":2048,"name":"generateWorkerOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateworkeroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":151,"kind":2048,"name":"generateLorryBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatelorrybody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":152,"kind":2048,"name":"generateLorryOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatelorryoptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":153,"kind":2048,"name":"generatePowerUpgraderBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatepowerupgraderbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":154,"kind":2048,"name":"generatePowerUpgraderOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatepowerupgraderoptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":155,"kind":2048,"name":"generateRemoteMinerBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremoteminerbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":156,"kind":2048,"name":"generateRemoteMinerOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremotemineroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":157,"kind":2048,"name":"generateRemoteHarvesterBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremoteharvesterbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":158,"kind":2048,"name":"generateRemoteHarvesterOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremoteharvesteroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":159,"kind":2048,"name":"generateRemoteReserverBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremotereserverbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":160,"kind":2048,"name":"generateRemoteReserverOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremotereserveroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":161,"kind":2048,"name":"generateRemoteColonizerBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremotecolonizerbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":162,"kind":2048,"name":"generateRemoteColonizerOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremotecolonizeroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":163,"kind":2048,"name":"generateRemoteDefenderBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremotedefenderbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":164,"kind":2048,"name":"generateRemoteDefenderOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generateremotedefenderoptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":165,"kind":2048,"name":"generateZealotBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatezealotbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":166,"kind":2048,"name":"generateZealotOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatezealotoptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":167,"kind":2048,"name":"generateMedicBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatemedicbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":168,"kind":2048,"name":"generateMedicOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatemedicoptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":169,"kind":2048,"name":"generateStalkerBody","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatestalkerbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":170,"kind":2048,"name":"generateStalkerOptions","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatestalkeroptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":171,"kind":2048,"name":"getDefaultCreepOptionsCiv","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#getdefaultcreepoptionsciv","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":172,"kind":2048,"name":"getDefaultCreepOptionsMili","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#getdefaultcreepoptionsmili","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":173,"kind":2048,"name":"generateDefaultCreepMemory","url":"classes/_src_helpers_spawnhelper_.spawnhelper.html#generatedefaultcreepmemory","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Helpers/SpawnHelper\".SpawnHelper"},{"id":174,"kind":1,"name":"\"src/Api/Spawn.Api\"","url":"modules/_src_api_spawn_api_.html","classes":"tsd-kind-external-module"},{"id":175,"kind":128,"name":"SpawnApi","url":"classes/_src_api_spawn_api_.spawnapi.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Api/Spawn.Api\""},{"id":176,"kind":2048,"name":"getCreepCount","url":"classes/_src_api_spawn_api_.spawnapi.html#getcreepcount","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":177,"kind":2048,"name":"getCreepLimits","url":"classes/_src_api_spawn_api_.spawnapi.html#getcreeplimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":178,"kind":2048,"name":"generateDomesticCreepLimits","url":"classes/_src_api_spawn_api_.spawnapi.html#generatedomesticcreeplimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":179,"kind":2048,"name":"generateRemoteCreepLimits","url":"classes/_src_api_spawn_api_.spawnapi.html#generateremotecreeplimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":180,"kind":2048,"name":"generateMilitaryCreepLimits","url":"classes/_src_api_spawn_api_.spawnapi.html#generatemilitarycreeplimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":181,"kind":2048,"name":"setCreepLimits","url":"classes/_src_api_spawn_api_.spawnapi.html#setcreeplimits","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":182,"kind":2048,"name":"getOpenSpawn","url":"classes/_src_api_spawn_api_.spawnapi.html#getopenspawn","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":183,"kind":2048,"name":"getNextCreep","url":"classes/_src_api_spawn_api_.spawnapi.html#getnextcreep","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":184,"kind":2048,"name":"spawnNextCreep","url":"classes/_src_api_spawn_api_.spawnapi.html#spawnnextcreep","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":185,"kind":2048,"name":"getEnergyCostOfBody","url":"classes/_src_api_spawn_api_.spawnapi.html#getenergycostofbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":186,"kind":2048,"name":"getTier","url":"classes/_src_api_spawn_api_.spawnapi.html#gettier","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":187,"kind":2048,"name":"generateCreepOptions","url":"classes/_src_api_spawn_api_.spawnapi.html#generatecreepoptions","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-private tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":188,"kind":2048,"name":"generateCreepBody","url":"classes/_src_api_spawn_api_.spawnapi.html#generatecreepbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":189,"kind":2048,"name":"getCreepBody","url":"classes/_src_api_spawn_api_.spawnapi.html#getcreepbody","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":190,"kind":2048,"name":"getActiveMiners","url":"classes/_src_api_spawn_api_.spawnapi.html#getactiveminers","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-private tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":191,"kind":2048,"name":"generateSquadUUID","url":"classes/_src_api_spawn_api_.spawnapi.html#generatesquaduuid","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"src/Api/Spawn.Api\".SpawnApi"},{"id":192,"kind":1,"name":"\"src/Helpers/CreepAllHelper\"","url":"modules/_src_helpers_CreepAllHelper_.html","classes":"tsd-kind-external-module"},{"id":193,"kind":128,"name":"CreepAllHelper","url":"classes/_src_helpers_CreepAllHelper_.CreepAllHelper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/CreepAllHelper\""},{"id":194,"kind":1,"name":"\"src/Helpers/StructureHelper\"","url":"modules/_src_helpers_structurehelper_.html","classes":"tsd-kind-external-module"},{"id":195,"kind":128,"name":"StructureHelper","url":"classes/_src_helpers_structurehelper_.structurehelper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"src/Helpers/StructureHelper\""},{"id":196,"kind":1,"name":"\"src/utils/StringMapApi\"","url":"modules/_src_utils_stringmapapi_.html","classes":"tsd-kind-external-module"},{"id":197,"kind":1,"name":"\"test/integration/helper\"","url":"modules/_test_integration_helper_.html","classes":"tsd-kind-external-module"},{"id":198,"kind":128,"name":"IntegrationTestHelper","url":"classes/_test_integration_helper_.integrationtesthelper.html","classes":"tsd-kind-class tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"test/integration/helper\""},{"id":199,"kind":1024,"name":"_server","url":"classes/_test_integration_helper_.integrationtesthelper.html#_server","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-private tsd-is-not-exported","parent":"\"test/integration/helper\".IntegrationTestHelper"},{"id":200,"kind":1024,"name":"_player","url":"classes/_test_integration_helper_.integrationtesthelper.html#_player","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-private tsd-is-not-exported","parent":"\"test/integration/helper\".IntegrationTestHelper"},{"id":201,"kind":262144,"name":"server","url":"classes/_test_integration_helper_.integrationtesthelper.html#server","classes":"tsd-kind-get-signature tsd-parent-kind-class tsd-is-not-exported","parent":"\"test/integration/helper\".IntegrationTestHelper"},{"id":202,"kind":262144,"name":"player","url":"classes/_test_integration_helper_.integrationtesthelper.html#player","classes":"tsd-kind-get-signature tsd-parent-kind-class tsd-is-not-exported","parent":"\"test/integration/helper\".IntegrationTestHelper"},{"id":203,"kind":2048,"name":"beforeEach","url":"classes/_test_integration_helper_.integrationtesthelper.html#beforeeach","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-not-exported","parent":"\"test/integration/helper\".IntegrationTestHelper"},{"id":204,"kind":2048,"name":"afterEach","url":"classes/_test_integration_helper_.integrationtesthelper.html#aftereach","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-not-exported","parent":"\"test/integration/helper\".IntegrationTestHelper"},{"id":205,"kind":32,"name":"readFileSync","url":"modules/_test_integration_helper_.html#readfilesync","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"test/integration/helper\""},{"id":206,"kind":32,"name":"_","url":"modules/_test_integration_helper_.html#_","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"test/integration/helper\""},{"id":207,"kind":32,"name":"ScreepsServer","url":"modules/_test_integration_helper_.html#screepsserver","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"test/integration/helper\""},{"id":208,"kind":32,"name":"stdHooks","url":"modules/_test_integration_helper_.html#stdhooks","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"test/integration/helper\""},{"id":209,"kind":32,"name":"DIST_MAIN_JS","url":"modules/_test_integration_helper_.html#dist_main_js","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"test/integration/helper\""},{"id":210,"kind":32,"name":"helper","url":"modules/_test_integration_helper_.html#helper","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/integration/helper\""},{"id":211,"kind":1,"name":"\"test/integration/SpawnApi.test\"","url":"modules/_test_integration_spawnapi_test_.html","classes":"tsd-kind-external-module"},{"id":212,"kind":1,"name":"\"test/integration/integration.test\"","url":"modules/_test_integration_integration_test_.html","classes":"tsd-kind-external-module"},{"id":213,"kind":1,"name":"\"test/unit/mock\"","url":"modules/_test_unit_mock_.html","classes":"tsd-kind-external-module"},{"id":214,"kind":2097152,"name":"Game","url":"modules/_test_unit_mock_.html#game","classes":"tsd-kind-object-literal tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":215,"kind":32,"name":"creeps","url":"modules/_test_unit_mock_.html#game.creeps","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".Game"},{"id":216,"kind":32,"name":"rooms","url":"modules/_test_unit_mock_.html#game.rooms","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".Game"},{"id":217,"kind":32,"name":"spawns","url":"modules/_test_unit_mock_.html#game.spawns","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".Game"},{"id":218,"kind":65536,"name":"__type","url":"modules/_test_unit_mock_.html#game.spawns.__type","classes":"tsd-kind-type-literal tsd-parent-kind-variable tsd-is-not-exported","parent":"\"test/unit/mock\".Game.spawns"},{"id":219,"kind":32,"name":"time","url":"modules/_test_unit_mock_.html#game.time","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".Game"},{"id":220,"kind":2097152,"name":"Memory","url":"modules/_test_unit_mock_.html#memory","classes":"tsd-kind-object-literal tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":221,"kind":32,"name":"creeps","url":"modules/_test_unit_mock_.html#memory.creeps-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".Memory"},{"id":222,"kind":32,"name":"OK","url":"modules/_test_unit_mock_.html#ok","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":223,"kind":32,"name":"ERR_NOT_OWNER","url":"modules/_test_unit_mock_.html#err_not_owner","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":224,"kind":32,"name":"ERR_NO_PATH","url":"modules/_test_unit_mock_.html#err_no_path","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":225,"kind":32,"name":"ERR_NAME_EXISTS","url":"modules/_test_unit_mock_.html#err_name_exists","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":226,"kind":32,"name":"ERR_BUSY","url":"modules/_test_unit_mock_.html#err_busy","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":227,"kind":32,"name":"ERR_NOT_FOUND","url":"modules/_test_unit_mock_.html#err_not_found","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":228,"kind":32,"name":"ERR_NOT_ENOUGH_RESOURCES","url":"modules/_test_unit_mock_.html#err_not_enough_resources","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":229,"kind":32,"name":"ERR_NOT_ENOUGH_ENERGY","url":"modules/_test_unit_mock_.html#err_not_enough_energy","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":230,"kind":32,"name":"ERR_INVALID_TARGET","url":"modules/_test_unit_mock_.html#err_invalid_target","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":231,"kind":32,"name":"ERR_FULL","url":"modules/_test_unit_mock_.html#err_full","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":232,"kind":32,"name":"ERR_NOT_IN_RANGE","url":"modules/_test_unit_mock_.html#err_not_in_range","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":233,"kind":32,"name":"ERR_INVALID_ARGS","url":"modules/_test_unit_mock_.html#err_invalid_args","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":234,"kind":32,"name":"ERR_TIRED","url":"modules/_test_unit_mock_.html#err_tired","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":235,"kind":32,"name":"ERR_NO_BODYPART","url":"modules/_test_unit_mock_.html#err_no_bodypart","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":236,"kind":32,"name":"ERR_NOT_ENOUGH_EXTENSIONS","url":"modules/_test_unit_mock_.html#err_not_enough_extensions","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":237,"kind":32,"name":"ERR_RCL_NOT_ENOUGH","url":"modules/_test_unit_mock_.html#err_rcl_not_enough","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":238,"kind":32,"name":"ERR_GCL_NOT_ENOUGH","url":"modules/_test_unit_mock_.html#err_gcl_not_enough","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":239,"kind":32,"name":"FIND_EXIT_TOP","url":"modules/_test_unit_mock_.html#find_exit_top","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":240,"kind":32,"name":"FIND_EXIT_RIGHT","url":"modules/_test_unit_mock_.html#find_exit_right","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":241,"kind":32,"name":"FIND_EXIT_BOTTOM","url":"modules/_test_unit_mock_.html#find_exit_bottom","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":242,"kind":32,"name":"FIND_EXIT_LEFT","url":"modules/_test_unit_mock_.html#find_exit_left","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":243,"kind":32,"name":"FIND_EXIT","url":"modules/_test_unit_mock_.html#find_exit","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":244,"kind":32,"name":"FIND_CREEPS","url":"modules/_test_unit_mock_.html#find_creeps","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":245,"kind":32,"name":"FIND_MY_CREEPS","url":"modules/_test_unit_mock_.html#find_my_creeps","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":246,"kind":32,"name":"FIND_HOSTILE_CREEPS","url":"modules/_test_unit_mock_.html#find_hostile_creeps","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":247,"kind":32,"name":"FIND_SOURCES_ACTIVE","url":"modules/_test_unit_mock_.html#find_sources_active","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":248,"kind":32,"name":"FIND_SOURCES","url":"modules/_test_unit_mock_.html#find_sources","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":249,"kind":32,"name":"FIND_DROPPED_ENERGY","url":"modules/_test_unit_mock_.html#find_dropped_energy","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":250,"kind":32,"name":"FIND_DROPPED_RESOURCES","url":"modules/_test_unit_mock_.html#find_dropped_resources","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":251,"kind":32,"name":"FIND_STRUCTURES","url":"modules/_test_unit_mock_.html#find_structures","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":252,"kind":32,"name":"FIND_MY_STRUCTURES","url":"modules/_test_unit_mock_.html#find_my_structures","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":253,"kind":32,"name":"FIND_HOSTILE_STRUCTURES","url":"modules/_test_unit_mock_.html#find_hostile_structures","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":254,"kind":32,"name":"FIND_FLAGS","url":"modules/_test_unit_mock_.html#find_flags","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":255,"kind":32,"name":"FIND_CONSTRUCTION_SITES","url":"modules/_test_unit_mock_.html#find_construction_sites","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":256,"kind":32,"name":"FIND_MY_SPAWNS","url":"modules/_test_unit_mock_.html#find_my_spawns","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":257,"kind":32,"name":"FIND_HOSTILE_SPAWNS","url":"modules/_test_unit_mock_.html#find_hostile_spawns","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":258,"kind":32,"name":"FIND_MY_CONSTRUCTION_SITES","url":"modules/_test_unit_mock_.html#find_my_construction_sites","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":259,"kind":32,"name":"FIND_HOSTILE_CONSTRUCTION_SITES","url":"modules/_test_unit_mock_.html#find_hostile_construction_sites","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":260,"kind":32,"name":"FIND_MINERALS","url":"modules/_test_unit_mock_.html#find_minerals","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":261,"kind":32,"name":"FIND_NUKES","url":"modules/_test_unit_mock_.html#find_nukes","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":262,"kind":32,"name":"FIND_TOMBSTONES","url":"modules/_test_unit_mock_.html#find_tombstones","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":263,"kind":32,"name":"TOP","url":"modules/_test_unit_mock_.html#top","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":264,"kind":32,"name":"TOP_RIGHT","url":"modules/_test_unit_mock_.html#top_right","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":265,"kind":32,"name":"RIGHT","url":"modules/_test_unit_mock_.html#right","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":266,"kind":32,"name":"BOTTOM_RIGHT","url":"modules/_test_unit_mock_.html#bottom_right","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":267,"kind":32,"name":"BOTTOM","url":"modules/_test_unit_mock_.html#bottom","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":268,"kind":32,"name":"BOTTOM_LEFT","url":"modules/_test_unit_mock_.html#bottom_left","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":269,"kind":32,"name":"LEFT","url":"modules/_test_unit_mock_.html#left","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":270,"kind":32,"name":"TOP_LEFT","url":"modules/_test_unit_mock_.html#top_left","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":271,"kind":32,"name":"COLOR_RED","url":"modules/_test_unit_mock_.html#color_red","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":272,"kind":32,"name":"COLOR_PURPLE","url":"modules/_test_unit_mock_.html#color_purple","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":273,"kind":32,"name":"COLOR_BLUE","url":"modules/_test_unit_mock_.html#color_blue","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":274,"kind":32,"name":"COLOR_CYAN","url":"modules/_test_unit_mock_.html#color_cyan","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":275,"kind":32,"name":"COLOR_GREEN","url":"modules/_test_unit_mock_.html#color_green","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":276,"kind":32,"name":"COLOR_YELLOW","url":"modules/_test_unit_mock_.html#color_yellow","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":277,"kind":32,"name":"COLOR_ORANGE","url":"modules/_test_unit_mock_.html#color_orange","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":278,"kind":32,"name":"COLOR_BROWN","url":"modules/_test_unit_mock_.html#color_brown","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":279,"kind":32,"name":"COLOR_GREY","url":"modules/_test_unit_mock_.html#color_grey","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":280,"kind":32,"name":"COLOR_WHITE","url":"modules/_test_unit_mock_.html#color_white","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":281,"kind":32,"name":"CREEP_SPAWN_TIME","url":"modules/_test_unit_mock_.html#creep_spawn_time","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":282,"kind":32,"name":"CREEP_LIFE_TIME","url":"modules/_test_unit_mock_.html#creep_life_time","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":283,"kind":32,"name":"CREEP_CLAIM_LIFE_TIME","url":"modules/_test_unit_mock_.html#creep_claim_life_time","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":284,"kind":32,"name":"CREEP_CORPSE_RATE","url":"modules/_test_unit_mock_.html#creep_corpse_rate","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":285,"kind":32,"name":"OBSTACLE_OBJECT_TYPES","url":"modules/_test_unit_mock_.html#obstacle_object_types","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":286,"kind":32,"name":"ENERGY_REGEN_TIME","url":"modules/_test_unit_mock_.html#energy_regen_time","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":287,"kind":32,"name":"ENERGY_DECAY","url":"modules/_test_unit_mock_.html#energy_decay","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":288,"kind":32,"name":"REPAIR_COST","url":"modules/_test_unit_mock_.html#repair_cost","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":289,"kind":32,"name":"RAMPART_DECAY_AMOUNT","url":"modules/_test_unit_mock_.html#rampart_decay_amount","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":290,"kind":32,"name":"RAMPART_DECAY_TIME","url":"modules/_test_unit_mock_.html#rampart_decay_time","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":291,"kind":32,"name":"RAMPART_HITS","url":"modules/_test_unit_mock_.html#rampart_hits","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":292,"kind":2097152,"name":"RAMPART_HITS_MAX","url":"modules/_test_unit_mock_.html#rampart_hits_max","classes":"tsd-kind-object-literal tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":293,"kind":32,"name":"2","url":"modules/_test_unit_mock_.html#rampart_hits_max.2-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".RAMPART_HITS_MAX"},{"id":294,"kind":32,"name":"3","url":"modules/_test_unit_mock_.html#rampart_hits_max.3-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".RAMPART_HITS_MAX"},{"id":295,"kind":32,"name":"4","url":"modules/_test_unit_mock_.html#rampart_hits_max.4-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".RAMPART_HITS_MAX"},{"id":296,"kind":32,"name":"5","url":"modules/_test_unit_mock_.html#rampart_hits_max.5-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".RAMPART_HITS_MAX"},{"id":297,"kind":32,"name":"6","url":"modules/_test_unit_mock_.html#rampart_hits_max.6-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".RAMPART_HITS_MAX"},{"id":298,"kind":32,"name":"7","url":"modules/_test_unit_mock_.html#rampart_hits_max.7-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".RAMPART_HITS_MAX"},{"id":299,"kind":32,"name":"8","url":"modules/_test_unit_mock_.html#rampart_hits_max.8-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".RAMPART_HITS_MAX"},{"id":300,"kind":32,"name":"SPAWN_HITS","url":"modules/_test_unit_mock_.html#spawn_hits","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":301,"kind":32,"name":"SPAWN_ENERGY_START","url":"modules/_test_unit_mock_.html#spawn_energy_start","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":302,"kind":32,"name":"SPAWN_ENERGY_CAPACITY","url":"modules/_test_unit_mock_.html#spawn_energy_capacity","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":303,"kind":32,"name":"SOURCE_ENERGY_CAPACITY","url":"modules/_test_unit_mock_.html#source_energy_capacity","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":304,"kind":32,"name":"SOURCE_ENERGY_NEUTRAL_CAPACITY","url":"modules/_test_unit_mock_.html#source_energy_neutral_capacity","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":305,"kind":32,"name":"SOURCE_ENERGY_KEEPER_CAPACITY","url":"modules/_test_unit_mock_.html#source_energy_keeper_capacity","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":306,"kind":32,"name":"WALL_HITS","url":"modules/_test_unit_mock_.html#wall_hits","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":307,"kind":32,"name":"WALL_HITS_MAX","url":"modules/_test_unit_mock_.html#wall_hits_max","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":308,"kind":32,"name":"EXTENSION_HITS","url":"modules/_test_unit_mock_.html#extension_hits","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":309,"kind":2097152,"name":"EXTENSION_ENERGY_CAPACITY","url":"modules/_test_unit_mock_.html#extension_energy_capacity","classes":"tsd-kind-object-literal tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":310,"kind":32,"name":"0","url":"modules/_test_unit_mock_.html#extension_energy_capacity.0","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":311,"kind":32,"name":"1","url":"modules/_test_unit_mock_.html#extension_energy_capacity.1","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":312,"kind":32,"name":"2","url":"modules/_test_unit_mock_.html#extension_energy_capacity.2","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":313,"kind":32,"name":"3","url":"modules/_test_unit_mock_.html#extension_energy_capacity.3","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":314,"kind":32,"name":"4","url":"modules/_test_unit_mock_.html#extension_energy_capacity.4","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":315,"kind":32,"name":"5","url":"modules/_test_unit_mock_.html#extension_energy_capacity.5","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":316,"kind":32,"name":"6","url":"modules/_test_unit_mock_.html#extension_energy_capacity.6","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":317,"kind":32,"name":"7","url":"modules/_test_unit_mock_.html#extension_energy_capacity.7","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":318,"kind":32,"name":"8","url":"modules/_test_unit_mock_.html#extension_energy_capacity.8","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"test/unit/mock\".EXTENSION_ENERGY_CAPACITY"},{"id":319,"kind":32,"name":"ROAD_HITS","url":"modules/_test_unit_mock_.html#road_hits","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":320,"kind":32,"name":"ROAD_WEAROUT","url":"modules/_test_unit_mock_.html#road_wearout","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":321,"kind":32,"name":"ROAD_DECAY_AMOUNT","url":"modules/_test_unit_mock_.html#road_decay_amount","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":322,"kind":32,"name":"ROAD_DECAY_TIME","url":"modules/_test_unit_mock_.html#road_decay_time","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":323,"kind":32,"name":"LINK_HITS","url":"modules/_test_unit_mock_.html#link_hits","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":324,"kind":32,"name":"LINK_HITS_MAX","url":"modules/_test_unit_mock_.html#link_hits_max","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":325,"kind":32,"name":"LINK_CAPACITY","url":"modules/_test_unit_mock_.html#link_capacity","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":326,"kind":32,"name":"LINK_COOLDOWN","url":"modules/_test_unit_mock_.html#link_cooldown","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":327,"kind":32,"name":"LINK_LOSS_RATIO","url":"modules/_test_unit_mock_.html#link_loss_ratio","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":328,"kind":32,"name":"STORAGE_CAPACITY","url":"modules/_test_unit_mock_.html#storage_capacity","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":329,"kind":32,"name":"STORAGE_HITS","url":"modules/_test_unit_mock_.html#storage_hits","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":330,"kind":32,"name":"CARRY_CAPACITY","url":"modules/_test_unit_mock_.html#carry_capacity","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":331,"kind":32,"name":"HARVEST_POWER","url":"modules/_test_unit_mock_.html#harvest_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":332,"kind":32,"name":"HARVEST_MINERAL_POWER","url":"modules/_test_unit_mock_.html#harvest_mineral_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":333,"kind":32,"name":"REPAIR_POWER","url":"modules/_test_unit_mock_.html#repair_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":334,"kind":32,"name":"DISMANTLE_POWER","url":"modules/_test_unit_mock_.html#dismantle_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":335,"kind":32,"name":"BUILD_POWER","url":"modules/_test_unit_mock_.html#build_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":336,"kind":32,"name":"ATTACK_POWER","url":"modules/_test_unit_mock_.html#attack_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":337,"kind":32,"name":"UPGRADE_CONTROLLER_POWER","url":"modules/_test_unit_mock_.html#upgrade_controller_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":338,"kind":32,"name":"RANGED_ATTACK_POWER","url":"modules/_test_unit_mock_.html#ranged_attack_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":339,"kind":32,"name":"HEAL_POWER","url":"modules/_test_unit_mock_.html#heal_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":340,"kind":32,"name":"RANGED_HEAL_POWER","url":"modules/_test_unit_mock_.html#ranged_heal_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":341,"kind":32,"name":"DISMANTLE_COST","url":"modules/_test_unit_mock_.html#dismantle_cost","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":342,"kind":32,"name":"MOVE","url":"modules/_test_unit_mock_.html#move","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":343,"kind":32,"name":"WORK","url":"modules/_test_unit_mock_.html#work","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":344,"kind":32,"name":"CARRY","url":"modules/_test_unit_mock_.html#carry","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":345,"kind":32,"name":"ATTACK","url":"modules/_test_unit_mock_.html#attack","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":346,"kind":32,"name":"RANGED_ATTACK","url":"modules/_test_unit_mock_.html#ranged_attack","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":347,"kind":32,"name":"TOUGH","url":"modules/_test_unit_mock_.html#tough","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":348,"kind":32,"name":"HEAL","url":"modules/_test_unit_mock_.html#heal","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":349,"kind":32,"name":"CLAIM","url":"modules/_test_unit_mock_.html#claim","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":350,"kind":32,"name":"BODYPARTS_ALL","url":"modules/_test_unit_mock_.html#bodyparts_all","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":351,"kind":32,"name":"CONSTRUCTION_COST_ROAD_SWAMP_RATIO","url":"modules/_test_unit_mock_.html#construction_cost_road_swamp_ratio","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":352,"kind":32,"name":"STRUCTURE_EXTENSION","url":"modules/_test_unit_mock_.html#structure_extension","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":353,"kind":32,"name":"STRUCTURE_RAMPART","url":"modules/_test_unit_mock_.html#structure_rampart","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":354,"kind":32,"name":"STRUCTURE_ROAD","url":"modules/_test_unit_mock_.html#structure_road","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":355,"kind":32,"name":"STRUCTURE_SPAWN","url":"modules/_test_unit_mock_.html#structure_spawn","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":356,"kind":32,"name":"STRUCTURE_LINK","url":"modules/_test_unit_mock_.html#structure_link","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":357,"kind":32,"name":"STRUCTURE_WALL","url":"modules/_test_unit_mock_.html#structure_wall","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":358,"kind":32,"name":"STRUCTURE_KEEPER_LAIR","url":"modules/_test_unit_mock_.html#structure_keeper_lair","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":359,"kind":32,"name":"STRUCTURE_CONTROLLER","url":"modules/_test_unit_mock_.html#structure_controller","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":360,"kind":32,"name":"STRUCTURE_STORAGE","url":"modules/_test_unit_mock_.html#structure_storage","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":361,"kind":32,"name":"STRUCTURE_TOWER","url":"modules/_test_unit_mock_.html#structure_tower","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":362,"kind":32,"name":"STRUCTURE_OBSERVER","url":"modules/_test_unit_mock_.html#structure_observer","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":363,"kind":32,"name":"STRUCTURE_POWER_BANK","url":"modules/_test_unit_mock_.html#structure_power_bank","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":364,"kind":32,"name":"STRUCTURE_POWER_SPAWN","url":"modules/_test_unit_mock_.html#structure_power_spawn","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":365,"kind":32,"name":"STRUCTURE_EXTRACTOR","url":"modules/_test_unit_mock_.html#structure_extractor","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":366,"kind":32,"name":"STRUCTURE_LAB","url":"modules/_test_unit_mock_.html#structure_lab","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":367,"kind":32,"name":"STRUCTURE_TERMINAL","url":"modules/_test_unit_mock_.html#structure_terminal","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":368,"kind":32,"name":"STRUCTURE_CONTAINER","url":"modules/_test_unit_mock_.html#structure_container","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":369,"kind":32,"name":"STRUCTURE_NUKER","url":"modules/_test_unit_mock_.html#structure_nuker","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":370,"kind":32,"name":"STRUCTURE_PORTAL","url":"modules/_test_unit_mock_.html#structure_portal","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":371,"kind":32,"name":"RESOURCE_ENERGY","url":"modules/_test_unit_mock_.html#resource_energy","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":372,"kind":32,"name":"RESOURCE_POWER","url":"modules/_test_unit_mock_.html#resource_power","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":373,"kind":32,"name":"RESOURCE_UTRIUM","url":"modules/_test_unit_mock_.html#resource_utrium","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":374,"kind":32,"name":"RESOURCE_LEMERGIUM","url":"modules/_test_unit_mock_.html#resource_lemergium","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":375,"kind":32,"name":"RESOURCE_KEANIUM","url":"modules/_test_unit_mock_.html#resource_keanium","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":376,"kind":32,"name":"RESOURCE_GHODIUM","url":"modules/_test_unit_mock_.html#resource_ghodium","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":377,"kind":32,"name":"RESOURCE_ZYNTHIUM","url":"modules/_test_unit_mock_.html#resource_zynthium","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":378,"kind":32,"name":"RESOURCE_OXYGEN","url":"modules/_test_unit_mock_.html#resource_oxygen","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":379,"kind":32,"name":"RESOURCE_HYDROGEN","url":"modules/_test_unit_mock_.html#resource_hydrogen","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":380,"kind":32,"name":"RESOURCE_CATALYST","url":"modules/_test_unit_mock_.html#resource_catalyst","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":381,"kind":32,"name":"RESOURCE_HYDROXIDE","url":"modules/_test_unit_mock_.html#resource_hydroxide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":382,"kind":32,"name":"RESOURCE_ZYNTHIUM_KEANITE","url":"modules/_test_unit_mock_.html#resource_zynthium_keanite","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":383,"kind":32,"name":"RESOURCE_UTRIUM_LEMERGITE","url":"modules/_test_unit_mock_.html#resource_utrium_lemergite","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":384,"kind":32,"name":"RESOURCE_UTRIUM_HYDRIDE","url":"modules/_test_unit_mock_.html#resource_utrium_hydride","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":385,"kind":32,"name":"RESOURCE_UTRIUM_OXIDE","url":"modules/_test_unit_mock_.html#resource_utrium_oxide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":386,"kind":32,"name":"RESOURCE_KEANIUM_HYDRIDE","url":"modules/_test_unit_mock_.html#resource_keanium_hydride","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":387,"kind":32,"name":"RESOURCE_KEANIUM_OXIDE","url":"modules/_test_unit_mock_.html#resource_keanium_oxide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":388,"kind":32,"name":"RESOURCE_LEMERGIUM_HYDRIDE","url":"modules/_test_unit_mock_.html#resource_lemergium_hydride","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":389,"kind":32,"name":"RESOURCE_LEMERGIUM_OXIDE","url":"modules/_test_unit_mock_.html#resource_lemergium_oxide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":390,"kind":32,"name":"RESOURCE_ZYNTHIUM_HYDRIDE","url":"modules/_test_unit_mock_.html#resource_zynthium_hydride","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":391,"kind":32,"name":"RESOURCE_ZYNTHIUM_OXIDE","url":"modules/_test_unit_mock_.html#resource_zynthium_oxide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":392,"kind":32,"name":"RESOURCE_GHODIUM_HYDRIDE","url":"modules/_test_unit_mock_.html#resource_ghodium_hydride","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":393,"kind":32,"name":"RESOURCE_GHODIUM_OXIDE","url":"modules/_test_unit_mock_.html#resource_ghodium_oxide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":394,"kind":32,"name":"RESOURCE_UTRIUM_ACID","url":"modules/_test_unit_mock_.html#resource_utrium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":395,"kind":32,"name":"RESOURCE_UTRIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_utrium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":396,"kind":32,"name":"RESOURCE_KEANIUM_ACID","url":"modules/_test_unit_mock_.html#resource_keanium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":397,"kind":32,"name":"RESOURCE_KEANIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_keanium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":398,"kind":32,"name":"RESOURCE_LEMERGIUM_ACID","url":"modules/_test_unit_mock_.html#resource_lemergium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":399,"kind":32,"name":"RESOURCE_LEMERGIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_lemergium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":400,"kind":32,"name":"RESOURCE_ZYNTHIUM_ACID","url":"modules/_test_unit_mock_.html#resource_zynthium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":401,"kind":32,"name":"RESOURCE_ZYNTHIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_zynthium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":402,"kind":32,"name":"RESOURCE_GHODIUM_ACID","url":"modules/_test_unit_mock_.html#resource_ghodium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":403,"kind":32,"name":"RESOURCE_GHODIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_ghodium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":404,"kind":32,"name":"RESOURCE_CATALYZED_UTRIUM_ACID","url":"modules/_test_unit_mock_.html#resource_catalyzed_utrium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":405,"kind":32,"name":"RESOURCE_CATALYZED_UTRIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_catalyzed_utrium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":406,"kind":32,"name":"RESOURCE_CATALYZED_KEANIUM_ACID","url":"modules/_test_unit_mock_.html#resource_catalyzed_keanium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":407,"kind":32,"name":"RESOURCE_CATALYZED_KEANIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_catalyzed_keanium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":408,"kind":32,"name":"RESOURCE_CATALYZED_LEMERGIUM_ACID","url":"modules/_test_unit_mock_.html#resource_catalyzed_lemergium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":409,"kind":32,"name":"RESOURCE_CATALYZED_LEMERGIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_catalyzed_lemergium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":410,"kind":32,"name":"RESOURCE_CATALYZED_ZYNTHIUM_ACID","url":"modules/_test_unit_mock_.html#resource_catalyzed_zynthium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":411,"kind":32,"name":"RESOURCE_CATALYZED_ZYNTHIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_catalyzed_zynthium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":412,"kind":32,"name":"RESOURCE_CATALYZED_GHODIUM_ACID","url":"modules/_test_unit_mock_.html#resource_catalyzed_ghodium_acid","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":413,"kind":32,"name":"RESOURCE_CATALYZED_GHODIUM_ALKALIDE","url":"modules/_test_unit_mock_.html#resource_catalyzed_ghodium_alkalide","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":414,"kind":32,"name":"SUBSCRIPTION_TOKEN","url":"modules/_test_unit_mock_.html#subscription_token","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":415,"kind":32,"name":"SAFE_MODE_DURATION","url":"modules/_test_unit_mock_.html#safe_mode_duration","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":416,"kind":32,"name":"SAFE_MODE_COOLDOWN","url":"modules/_test_unit_mock_.html#safe_mode_cooldown","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":417,"kind":32,"name":"SAFE_MODE_COST","url":"modules/_test_unit_mock_.html#safe_mode_cost","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":418,"kind":32,"name":"LOOK_CREEPS","url":"modules/_test_unit_mock_.html#look_creeps","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":419,"kind":32,"name":"LOOK_ENERGY","url":"modules/_test_unit_mock_.html#look_energy","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":420,"kind":32,"name":"LOOK_RESOURCES","url":"modules/_test_unit_mock_.html#look_resources","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":421,"kind":32,"name":"LOOK_SOURCES","url":"modules/_test_unit_mock_.html#look_sources","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":422,"kind":32,"name":"LOOK_MINERALS","url":"modules/_test_unit_mock_.html#look_minerals","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":423,"kind":32,"name":"LOOK_STRUCTURES","url":"modules/_test_unit_mock_.html#look_structures","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":424,"kind":32,"name":"LOOK_FLAGS","url":"modules/_test_unit_mock_.html#look_flags","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":425,"kind":32,"name":"LOOK_CONSTRUCTION_SITES","url":"modules/_test_unit_mock_.html#look_construction_sites","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":426,"kind":32,"name":"LOOK_NUKES","url":"modules/_test_unit_mock_.html#look_nukes","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":427,"kind":32,"name":"LOOK_TERRAIN","url":"modules/_test_unit_mock_.html#look_terrain","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":428,"kind":32,"name":"LOOK_TOMBSTONES","url":"modules/_test_unit_mock_.html#look_tombstones","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":429,"kind":32,"name":"ORDER_SELL","url":"modules/_test_unit_mock_.html#order_sell","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":430,"kind":32,"name":"ORDER_BUY","url":"modules/_test_unit_mock_.html#order_buy","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":431,"kind":32,"name":"TOMBSTONE_DECAY_PER_PART","url":"modules/_test_unit_mock_.html#tombstone_decay_per_part","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":432,"kind":32,"name":"EVENT_ATTACK","url":"modules/_test_unit_mock_.html#event_attack","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":433,"kind":32,"name":"EVENT_OBJECT_DESTROYED","url":"modules/_test_unit_mock_.html#event_object_destroyed","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":434,"kind":32,"name":"EVENT_ATTACK_CONTROLLER","url":"modules/_test_unit_mock_.html#event_attack_controller","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":435,"kind":32,"name":"EVENT_BUILD","url":"modules/_test_unit_mock_.html#event_build","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":436,"kind":32,"name":"EVENT_HARVEST","url":"modules/_test_unit_mock_.html#event_harvest","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":437,"kind":32,"name":"EVENT_HEAL","url":"modules/_test_unit_mock_.html#event_heal","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":438,"kind":32,"name":"EVENT_REPAIR","url":"modules/_test_unit_mock_.html#event_repair","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":439,"kind":32,"name":"EVENT_RESERVE_CONTROLLER","url":"modules/_test_unit_mock_.html#event_reserve_controller","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":440,"kind":32,"name":"EVENT_UPGRADE_CONTROLLER","url":"modules/_test_unit_mock_.html#event_upgrade_controller","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":441,"kind":32,"name":"EVENT_EXIT","url":"modules/_test_unit_mock_.html#event_exit","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":442,"kind":32,"name":"EVENT_ATTACK_TYPE_MELEE","url":"modules/_test_unit_mock_.html#event_attack_type_melee","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":443,"kind":32,"name":"EVENT_ATTACK_TYPE_RANGED","url":"modules/_test_unit_mock_.html#event_attack_type_ranged","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":444,"kind":32,"name":"EVENT_ATTACK_TYPE_RANGED_MASS","url":"modules/_test_unit_mock_.html#event_attack_type_ranged_mass","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":445,"kind":32,"name":"EVENT_ATTACK_TYPE_DISMANTLE","url":"modules/_test_unit_mock_.html#event_attack_type_dismantle","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":446,"kind":32,"name":"EVENT_ATTACK_TYPE_HIT_BACK","url":"modules/_test_unit_mock_.html#event_attack_type_hit_back","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":447,"kind":32,"name":"EVENT_ATTACK_TYPE_NUKE","url":"modules/_test_unit_mock_.html#event_attack_type_nuke","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":448,"kind":32,"name":"EVENT_HEAL_TYPE_MELEE","url":"modules/_test_unit_mock_.html#event_heal_type_melee","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":449,"kind":32,"name":"EVENT_HEAL_TYPE_RANGED","url":"modules/_test_unit_mock_.html#event_heal_type_ranged","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"test/unit/mock\""},{"id":450,"kind":1,"name":"\"test/unit/SpawnApi.test\"","url":"modules/_test_unit_spawnapi_test_.html","classes":"tsd-kind-external-module"},{"id":451,"kind":1,"name":"\"test/unit/main.test\"","url":"modules/_test_unit_main_test_.html","classes":"tsd-kind-external-module"}]};
