+++ docs/classes/_src_helpers_creephelper_.creephelper.html
@@ -3,7 +3,7 @@
 <head>
 	<meta charset="utf-8">
 	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<title>CreepHelper | screeps-typescript-starter</title>
+	<title>CreepAllHelper | screeps-typescript-starter</title>
 	<meta name="description" content="">
 	<meta name="viewport" content="width=device-width, initial-scale=1">
 	<link rel="stylesheet" href="../assets/css/main.css">
@@ -56,13 +56,13 @@
 					<a href="../globals.html">Globals</a>
 				</li>
 				<li>
-					<a href="../modules/_src_helpers_creephelper_.html">&quot;src/Helpers/CreepHelper&quot;</a>
+					<a href="../modules/_src_helpers_CreepAllHelper_.html">&quot;src/Helpers/CreepAllHelper&quot;</a>
 				</li>
 				<li>
-					<a href="_src_helpers_creephelper_.creephelper.html">CreepHelper</a>
+					<a href="_src_helpers_CreepAllHelper_.CreepAllHelper.html">CreepAllHelper</a>
 				</li>
 			</ul>
-			<h1>Class CreepHelper</h1>
+			<h1>Class CreepAllHelper</h1>
 		</div>
 	</div>
 </header>
@@ -73,7 +73,7 @@
 				<h3>Hierarchy</h3>
 				<ul class="tsd-hierarchy">
 					<li>
-						<span class="target">CreepHelper</span>
+						<span class="target">CreepAllHelper</span>
 					</li>
 				</ul>
 			</section>
@@ -85,7 +85,7 @@
 						<a href="../globals.html"><em>Globals</em></a>
 					</li>
 					<li class="current tsd-kind-external-module">
-						<a href="../modules/_src_helpers_creephelper_.html">"src/<wbr>Helpers/<wbr>Creep<wbr>Helper"</a>
+						<a href="../modules/_src_helpers_CreepAllHelper_.html">"src/<wbr>Helpers/<wbr>Creep<wbr>Helper"</a>
 					</li>
 				</ul>
 			</nav>
@@ -94,7 +94,7 @@
 				</ul>
 				<ul class="current">
 					<li class="current tsd-kind-class tsd-parent-kind-external-module">
-						<a href="_src_helpers_creephelper_.creephelper.html" class="tsd-kind-icon">Creep<wbr>Helper</a>
+						<a href="_src_helpers_CreepAllHelper_.CreepAllHelper.html" class="tsd-kind-icon">Creep<wbr>Helper</a>
 					</li>
 				</ul>
 				<ul class="after-current">
