+++ docs/globals.html
@@ -78,7 +78,7 @@
 								<li class="tsd-kind-external-module"><a href="modules/_src_api_room_api_.html" class="tsd-kind-icon">"src/<wbr>Api/<wbr>Room.<wbr>Api"</a></li>
 								<li class="tsd-kind-external-module"><a href="modules/_src_api_spawn_api_.html" class="tsd-kind-icon">"src/<wbr>Api/<wbr>Spawn.<wbr>Api"</a></li>
 								<li class="tsd-kind-external-module"><a href="modules/_src_helpers_consolecommands_.html" class="tsd-kind-icon">"src/<wbr>Helpers/<wbr>Console<wbr>Commands"</a></li>
-								<li class="tsd-kind-external-module"><a href="modules/_src_helpers_creephelper_.html" class="tsd-kind-icon">"src/<wbr>Helpers/<wbr>Creep<wbr>Helper"</a></li>
+								<li class="tsd-kind-external-module"><a href="modules/_src_helpers_CreepAllHelper_.html" class="tsd-kind-icon">"src/<wbr>Helpers/<wbr>Creep<wbr>Helper"</a></li>
 								<li class="tsd-kind-external-module"><a href="modules/_src_helpers_memoryhelper_.html" class="tsd-kind-icon">"src/<wbr>Helpers/<wbr>Memory<wbr>Helper"</a></li>
 								<li class="tsd-kind-external-module"><a href="modules/_src_helpers_memoryhelper_room_.html" class="tsd-kind-icon">"src/<wbr>Helpers/<wbr>Memory<wbr>Helper_<wbr><wbr>Room"</a></li>
 								<li class="tsd-kind-external-module"><a href="modules/_src_helpers_normalize_.html" class="tsd-kind-icon">"src/<wbr>Helpers/<wbr>Normalize"</a></li>
@@ -141,7 +141,7 @@
 						<a href="modules/_src_helpers_consolecommands_.html">"src/<wbr>Helpers/<wbr>Console<wbr>Commands"</a>
 					</li>
 					<li class=" tsd-kind-external-module">
-						<a href="modules/_src_helpers_creephelper_.html">"src/<wbr>Helpers/<wbr>Creep<wbr>Helper"</a>
+						<a href="modules/_src_helpers_CreepAllHelper_.html">"src/<wbr>Helpers/<wbr>Creep<wbr>Helper"</a>
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="modules/_src_helpers_memoryhelper_.html">"src/<wbr>Helpers/<wbr>Memory<wbr>Helper"</a>
