+++ docs/index.html
@@ -132,7 +132,7 @@ yarn</code></pre>
 						<a href="modules/_src_helpers_consolecommands_.html">"src/<wbr>Helpers/<wbr>Console<wbr>Commands"</a>
 					</li>
 					<li class=" tsd-kind-external-module">
-						<a href="modules/_src_helpers_creephelper_.html">"src/<wbr>Helpers/<wbr>Creep<wbr>Helper"</a>
+						<a href="modules/_src_helpers_CreepAllHelper_.html">"src/<wbr>Helpers/<wbr>Creep<wbr>Helper"</a>
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="modules/_src_helpers_memoryhelper_.html">"src/<wbr>Helpers/<wbr>Memory<wbr>Helper"</a>
