+++ docs/modules/_src_helpers_creephelper_.html
@@ -3,7 +3,7 @@
 <head>
 	<meta charset="utf-8">
 	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<title>&quot;src/Helpers/CreepHelper&quot; | screeps-typescript-starter</title>
+	<title>&quot;src/Helpers/CreepAllHelper&quot; | screeps-typescript-starter</title>
 	<meta name="description" content="">
 	<meta name="viewport" content="width=device-width, initial-scale=1">
 	<link rel="stylesheet" href="../assets/css/main.css">
@@ -56,10 +56,10 @@
 					<a href="../globals.html">Globals</a>
 				</li>
 				<li>
-					<a href="_src_helpers_creephelper_.html">&quot;src/Helpers/CreepHelper&quot;</a>
+					<a href="_src_helpers_CreepAllHelper_.html">&quot;src/Helpers/CreepAllHelper&quot;</a>
 				</li>
 			</ul>
-			<h1>External module &quot;src/Helpers/CreepHelper&quot;</h1>
+			<h1>External module &quot;src/Helpers/CreepAllHelper&quot;</h1>
 		</div>
 	</div>
 </header>
@@ -73,7 +73,7 @@
 						<section class="tsd-index-section ">
 							<h3>Classes</h3>
 							<ul class="tsd-index-list">
-								<li class="tsd-kind-class tsd-parent-kind-external-module"><a href="../classes/_src_helpers_creephelper_.creephelper.html" class="tsd-kind-icon">Creep<wbr>Helper</a></li>
+								<li class="tsd-kind-class tsd-parent-kind-external-module"><a href="../classes/_src_helpers_CreepAllHelper_.CreepAllHelper.html" class="tsd-kind-icon">Creep<wbr>Helper</a></li>
 							</ul>
 						</section>
 					</div>
@@ -87,14 +87,14 @@
 						<a href="../globals.html"><em>Globals</em></a>
 					</li>
 					<li class="current tsd-kind-external-module">
-						<a href="_src_helpers_creephelper_.html">"src/<wbr>Helpers/<wbr>Creep<wbr>Helper"</a>
+						<a href="_src_helpers_CreepAllHelper_.html">"src/<wbr>Helpers/<wbr>Creep<wbr>Helper"</a>
 					</li>
 				</ul>
 			</nav>
 			<nav class="tsd-navigation secondary menu-sticky">
 				<ul class="before-current">
 					<li class=" tsd-kind-class tsd-parent-kind-external-module">
-						<a href="../classes/_src_helpers_creephelper_.creephelper.html" class="tsd-kind-icon">Creep<wbr>Helper</a>
+						<a href="../classes/_src_helpers_CreepAllHelper_.CreepAllHelper.html" class="tsd-kind-icon">Creep<wbr>Helper</a>
 					</li>
 				</ul>
 			</nav>
