+++ src/Creeps/Creep.Civ.Helper.ts
@@ -1,0 +1,3 @@
+export class CreepCivHelper {
+
+}
