+++ src/Creeps/Roles/HarvesterCreepManager.ts
@@ -1,4 +1,4 @@
-import { ROLE_HARVESTER, MemoryApi, CreepApi, MemoryHelper, CreepHelper } from "Utils/Imports/internals";
+import { ROLE_HARVESTER, MemoryApi, CreepAllApi, MemoryHelper, CreepAllHelper } from "Utils/Imports/internals";
 
 // Manager for the miner creep role
 export class HarvesterCreepManager implements ICivCreepRoleManager {
@@ -19,10 +19,10 @@ export class HarvesterCreepManager implements ICivCreepRoleManager {
     public getNewJob(creep: Creep, room: Room): BaseJob | undefined {
         // if creep is empty, get a GetEnergyJob
         if (creep.carry.energy === 0) {
-            return CreepApi.newGetEnergyJob(creep, room);
+            return CreepAllApi.newGetEnergyJob(creep, room);
         } else {
             let job: BaseJob | undefined = this.newCarryPartJob(creep, room);
-            if (job === undefined && CreepHelper.bodyPartExists(creep, WORK)) {
+            if (job === undefined && CreepAllHelper.bodyPartExists(creep, WORK)) {
                 job = this.newWorkPartJob(creep, room);
             }
             return job;
