+++ src/Creeps/Roles/MinerCreepManager.ts
@@ -1,4 +1,4 @@
-import { ROLE_MINER, CreepHelper, CreepApi, MemoryApi } from "Utils/Imports/internals";
+import { ROLE_MINER, CreepAllHelper, CreepAllApi, MemoryApi } from "Utils/Imports/internals";
 
 // Manager for the miner creep role
 export class MinerCreepManager implements ICivCreepRoleManager {
@@ -16,7 +16,7 @@ export class MinerCreepManager implements ICivCreepRoleManager {
      * @param room
      */
     public getNewJob(creep: Creep, room: Room): BaseJob | undefined {
-        return CreepApi.getNewSourceJob(creep, room);
+        return CreepAllApi.getNewSourceJob(creep, room);
     }
 
     /**
@@ -29,7 +29,7 @@ export class MinerCreepManager implements ICivCreepRoleManager {
         MemoryApi.updateJobMemory(creep, room);
 
         const isSource: boolean = true;
-        const miningContainer = CreepHelper.getMiningContainer(
+        const miningContainer = CreepAllHelper.getMiningContainer(
             creep.memory.job as GetEnergyJob,
             Game.rooms[creep.memory.homeRoom],
             isSource
