+++ src/Creeps/Roles/StalkerCreepManager.ts
@@ -1,4 +1,4 @@
-import { ROLE_STALKER, MiliApi } from "Utils/Imports/internals";
+import { ROLE_STALKER, CreepMiliApi } from "Utils/Imports/internals";
 
 // Manager for the miner creep role
 export class StalkerCreepManager implements IMiliCreepRoleManager {
@@ -18,33 +18,33 @@ export class StalkerCreepManager implements IMiliCreepRoleManager {
         const CREEP_RANGE: number = 3;
 
         // Carry out the basics of a military creep before moving on to specific logic
-        if (MiliApi.checkMilitaryCreepBasics(creep, creepOptions)) {
+        if (CreepMiliApi.checkMilitaryCreepBasics(creep, creepOptions)) {
             return;
         }
 
         // Find a target for the creep
-        creepOptions.attackTarget = MiliApi.getAttackTarget(creep, creepOptions, CREEP_RANGE);
+        creepOptions.attackTarget = CreepMiliApi.getAttackTarget(creep, creepOptions, CREEP_RANGE);
         const target: Creep | Structure<StructureConstant> | undefined = creepOptions.attackTarget;
         const isMelee: boolean = false;
         if (!target) {
             // Keep the creeps together in the squad, if they're in a squad
             if (creepOptions.squadUUID) {
-                MiliApi.moveCreepToFurthestSquadMember(creep);
+                CreepMiliApi.moveCreepToFurthestSquadMember(creep);
             }
             return; // idle if no current target
         }
         // If we aren't in attack range, move towards the attack target
-        if (!MiliApi.isInAttackRange(creep, target.pos, isMelee)) {
+        if (!CreepMiliApi.isInAttackRange(creep, target.pos, isMelee)) {
             creep.moveTo(target);
             return;
         } else {
-            MiliApi.kiteEnemyCreep(creep);
+            CreepMiliApi.kiteEnemyCreep(creep);
         }
 
         // We are in attack range and healthy, attack the target
         creep.rangedAttack(target);
 
         // Reset offensive target
-        MiliApi.resetOffensiveTarget(creep);
+        CreepMiliApi.resetOffensiveTarget(creep);
     }
 }
