+++ src/Creeps/Roles/ZealotCreepManager.ts
@@ -1,4 +1,4 @@
-import { ROLE_ZEALOT, MiliApi } from "Utils/Imports/internals";
+import { ROLE_ZEALOT, CreepMiliApi } from "Utils/Imports/internals";
 
 // Manager for the miner creep role
 export class ZealotCreepManager implements IMiliCreepRoleManager {
@@ -18,23 +18,23 @@ export class ZealotCreepManager implements IMiliCreepRoleManager {
         const CREEP_RANGE: number = 1;
 
         // Carry out the basics of a military creep before moving on to specific logic
-        if (MiliApi.checkMilitaryCreepBasics(creep, creepOptions)) {
+        if (CreepMiliApi.checkMilitaryCreepBasics(creep, creepOptions)) {
             return;
         }
 
         // Find a target for the creep
-        creepOptions.attackTarget = MiliApi.getAttackTarget(creep, creepOptions, CREEP_RANGE);
+        creepOptions.attackTarget = CreepMiliApi.getAttackTarget(creep, creepOptions, CREEP_RANGE);
         const target: Creep | Structure<StructureConstant> | undefined = creepOptions.attackTarget;
         const isMelee: boolean = true;
         if (!target) {
             // Keep the creeps together in the squad, if they're in a squad
             if (creepOptions.squadUUID) {
-                MiliApi.moveCreepToFurthestSquadMember(creep);
+                CreepMiliApi.moveCreepToFurthestSquadMember(creep);
             }
             return; // idle if no current target
         }
         // If we aren't in attack range, move towards the attack target
-        if (!MiliApi.isInAttackRange(creep, target.pos, isMelee)) {
+        if (!CreepMiliApi.isInAttackRange(creep, target.pos, isMelee)) {
             creep.moveTo(target);
             return;
         }
@@ -42,6 +42,6 @@ export class ZealotCreepManager implements IMiliCreepRoleManager {
         // We are in attack range and healthy, attack the target
         creep.attack(target);
         // Reset creep's target
-        MiliApi.resetOffensiveTarget(creep);
+        CreepMiliApi.resetOffensiveTarget(creep);
     }
 }
