+++ src/RoomVisuals/RoomVisual.Api.ts
@@ -18,7 +18,7 @@ import {
     MemoryApi,
     RoomVisualHelper,
     RoomHelper,
-    CreepHelper,
+    CreepAllHelper,
     TOWER_MIN_DAMAGE_THRESHOLD,
     TOWER_MAX_DAMAGE_THRESHOLD
 } from "Utils/Imports/internals";
@@ -644,11 +644,11 @@ export class RoomVisualApi {
         }
 
         // Take out creeps with heal parts
-        const healCreeps = _.remove(hostileCreeps, (c: Creep) => CreepHelper.bodyPartExists(c, HEAL));
+        const healCreeps = _.remove(hostileCreeps, (c: Creep) => CreepAllHelper.bodyPartExists(c, HEAL));
 
         // Take out creeps that can attack
         const attackCreeps = _.remove(hostileCreeps, (c: Creep) =>
-            CreepHelper.bodyPartExists(c, ATTACK, RANGED_ATTACK)
+            CreepAllHelper.bodyPartExists(c, ATTACK, RANGED_ATTACK)
         );
 
         // rename for clarity, all creeps leftover should be civilian
