+++ src/Spawn/CreepLimitHelpers/AdvancedStateCreepLimits.ts
@@ -36,7 +36,7 @@ export class AdvancedStateCreepLimits implements ICreepSpawnLimits {
     // To recreate with jobs, make a folder for jobs, make an interface for each different target type or job type
     // implement a doWork and a travelTo for each of these
     // then replace the switch statement with the for loop search for the correct type, and call the doWork or travelTo on it
-    // The creep manager will still just have creepapi.doWork, but that function will contain the search for the class we want
+    // The creep manager will still just have CreepAllApi.doWork, but that function will contain the search for the class we want
 
     /**
      * generate the domestic limits for the room
